+++ public/index.html
@@ -1,0 +1,53 @@
+<!DOCTYPE html>
+<html lang="en">
+   <head>
+      <meta charset="utf-8">
+      <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+      <meta name="theme-color" content="#000000">
+      <title>Unxd Sdk playground</title>
+      <!--required frame works should be loaded here-->
+   <link href="app.css" rel="stylesheet"></head>
+
+   <body>
+       
+
+
+        <h1 class="main-head">SDK will be loaded here</h1>
+        <hr/>
+
+        <section class="test-block" >
+            <article>
+                <h2>Es6</h2>
+                <hr/>
+                <div id="es6Root">
+                    <div id="unbxdTxtWrap">
+                    <input id="unbxdInput"/>
+                    <button id="searchBtn">search</button>
+                    
+                </div>
+                <div style="display: flex">
+                    <div id="paginationContainer"></div>
+                    <div id="sortWrapper"></div>
+                    <div id="didYouMeanWrapper"></div>
+                    <div id="selectedFacetWrapper"></div>
+                    <div id="productViewTypeContainer"></div>
+                </div>
+                <div id="breadcrumpContainer"></div>
+                <div id="bannerContainer"></div>
+                <div style="display:flex">
+                    <div>
+                        <div id="bucketedFacetWrapper"></div>
+                        <div id="facetsWrapper"></div>
+                        <div id="rangeFacetWrapper"></div>
+                    </div>
+                    <div id="searchResultsWrapper"></div>
+                    <div id="loaderContainer"></div>
+                    <div id="noResultWrapper"></div>
+                </div>
+
+                </div>
+
+            </article>
+        </section>  
+   <script type="text/javascript" src="js/app.js"></script><script type="text/javascript" src="js/es6SDK.js"></script></body>
+</html>
