+++ public/js/app.js
@@ -1,0 +1,1 @@
+!function(t){var e={};function n(a){if(e[a])return e[a].exports;var i=e[a]={i:a,l:!1,exports:{}};return t[a].call(i.exports,i,i.exports,n),i.l=!0,i.exports}n.m=t,n.c=e,n.d=function(t,e,a){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:a})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var a=Object.create(null);if(n.r(a),Object.defineProperty(a,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var i in t)n.d(a,i,function(e){return t[e]}.bind(null,i));return a},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=2)}([function(t,e,n){"use strict";n.r(e);var a={siteKey:"api key",apiKey:"site key",isFlag:!0,sdkHostName:"https://search.unbxd.io/",productType:"SEARCH",searchQueryParam:"q",fields:["title"],defaultFilters:null,spellCheck:!1,pageSize:5,startPageNo:0,facetMultiSelect:!0,facetMultiSelectionMode:!0,searchQueryParam:null,updateUrls:!0,variants:!1,variantMapping:{},variantConfig:{count:1},extraParams:"",facetMultilevel:!0,facetDepth:6,productId:"uniqueId",swatches:!0,swatchMap:{},callBackFn:()=>{}};const i={extraParamUrlString:function(){const{extraParams:t}=this.options;let e="";return Object.keys(t).forEach(n=>{e+="&"+n+"="+t[n]}),e},renderFromUrl:function(){let t=!1;const e=this.getQueryParams();if("object"==typeof e){const{q:n,filter:a,sort:i,start:s}=e;t=!0,this.state.userInput!==n&&(this.state.userInput=n);const r=this.getFilterFromParams(a);this.state.rangeFacet=r.rangeFacet,this.state.selectedFacets=r.selectedFilters,this.state.selectedSort=i||"",this.state.startPageNo=s||0,this.state.categoryFilter=this.getCategoryFilterFromParams(e)}t&&this.getResults()},getQueryParams:function(){const t=decodeURI(location.hash.substr(1)).split("&");let e=[];t.forEach((t,n)=>{t.indexOf("=")>0?e.push(t):e[n-1]=e[n-1]+"&"+t});let n={};for(let t=0;t<e.length;t++){const a=e[t].split("=");void 0===n[a[0]]?n[a[0]]=a[1]:"string"==typeof n[a[0]]?(n[a[0]]=[n[a[0]]],n[a[0]].push(a[1])):n[a[0]].push(a[1])}return n},urlFlattenFacets:function(){let t="";const{selectedFacets:e}=this.state;return Object.keys(e).forEach(n=>{const a=e[n];t+="&filter=",a.forEach((e,a)=>{const{name:i}=e;t+=0===a?n+':"'+JSON.stringify(i).replace('"',""):" OR "+n+':"'+JSON.stringify(i).replace('"',"")})}),t},getDefaultFiltersStr:function(){let t="";const e=this.options.defaultFilters;if(null!==e){Object.keys(e).forEach(n=>{t+="&filter="+n+":"+JSON.stringify(e[n])})}return t},getFieldsStr:function(){return this.options.fields?"&fields="+this.options.fields.join(","):""},getNoOfResultsStr:function(){return"&rows="+this.options.pageSize},getPageStartStr:function(){return"&start="+this.state.startPageNo},getNewUrlState:function(){const t=this.state.userInput,e=this.urlFlattenFacets(),n=this.getFieldsStr()+this.getDefaultFiltersStr(),a=this.getSortUrlString(),i=this.getRangeFilterStr(),s=this.categoryFilterUrlStr();let r=this.options.facetMultiSelect?"&facet.multiselect=true":"";return this.state.currentUrl=this.url+t+s+r+e+n+a+i+this.getNoOfResultsStr()+this.getPageStartStr()+this.extraParamUrlString(),this.state.currentUrl},getRangeFilterStr:function(){const{rangeFacet:t}=this.state,e=Object.keys(t);let n="";return e.forEach(e=>{n+="&filter="+e+":"+t[e]}),n},categoryFilterUrlStr:function(){const{categoryFilter:t}=this.state;let e="";return Object.keys(t).forEach((n,a)=>{const i=t[n].map((t,e)=>{let n=encodeURIComponent(t);return e>0&&(n=">"+n),n});e+="&category-filter="+i.join("")}),e},getSortUrlString:function(){const t=this.getSelectedSort();return t?"&sort="+encodeURI(t):""},getBaseUrl:function(){const{sdkHostName:t,siteKey:e,apiKey:n}=this.options;return t+n+"/"+e}};var s=function(t){this.state.startPageNo=t},r=function(){const t=this.getSearchResults(),{numberOfProducts:e,start:n,products:a}=t,i=this.options.pageSize,s=Math.floor(e/i)+1;let r=Math.floor(n/i)+1;const c=!(n+i>=e||n>=e),o=!(n-i<0||n<=0);return{numberOfProducts:e,start:n,productsLn:a.length,rows:i,noOfPages:s,currentPage:r,isNext:c,isPrev:o}};var c=function(){return this.state.responseObj.searchMetaData||null},o=function(){const t=this.getResponseObj();if(t){const{response:e}=t;return e||null}return null},l=function(){const t=this.getSearchMeta();return t&&t.queryParams||null},u=function(){return this.getSearchQueryParams().q||""};var h={changeInput:"CHANGE_INPUT",fetchError:"FETCH_ERROR",afterApiCall:"AFTER_API_CALL",beforeApiCall:"BEFORE_API_CALL"};const d=t=>{let e=[];return(t||[]).forEach((t,n)=>{const{values:a}=t;t.values=((t=[])=>{let e=[];return t.forEach((n,a)=>{const i=n;let s="";a%2==1&&(s=t[a-1],e.push({name:s,count:i,dataId:a}))}),e})(a),e.push(t)}),e};var p=function(t=""){const{userInput:e}=this.state,n=""===t?e:t;this.state.userInput=n,this.state.currentUrl=this.getNewUrlState(),this.state.isLoading=!0,this.callBack(this,h.beforeApiCall);const a=fetch(this.state.currentUrl),i=this;a.then(t=>t.json()).then(t=>{if(this.state.isLoading=!1,t){this.options.variants&&(t=this.processVariantMap(t)),this.state.responseObj=t;const e=this.getDidYouMeanFromResponse();if(this.options.spellCheck&&e){this.state.didYouMean=e;if(this.getSpellCheckSuggested())return this.getResults(this.getSpellCheckSuggested()),!1}const n=t.facets;if(n){const e=this.getFacets(),a=d.bind(this)(e);t.facets={...n,text:{list:a}}}this.state.responseObj=t,this.setSort(),this.setStateFromData(),this.setUrl(),this.callBack(i,h.afterApiCall)}}).catch(t=>{console.log(this,t),this.state.isLoading=!1,this.callBack(i,h.fetchError)})},f=function(t={}){const e=this;if(t.response&&t.response.products){let n=t.response.products.map(t=>{const{variants:n=[],relevantDocument:a}=t;let i=t;if("variant"===a&&n.length>0){const t=n[0],a=e.options.variantMapping;Object.keys(a).forEach(e=>{const n=a[e];i[e]=t[n]})}return i});t.response.products=n}return t},g=function(t,e){const n=this.getSearchResults(),{products:a}=n;return a.find(n=>n[t]===e)},m=function(){const t=this.getSearchQueryParams(),{q:e,filter:n}=t;this.state.userInput=e;const a=this.getFilterFromParams(n);this.state.rangeFacet=a.rangeFacet,this.state.selectedFacets=a.selectedFilters,this.state.categoryFilter=this.getCategoryFilterFromParams(t),this.state.breadcrumbs=this.getBreadCrumbs()};var v=function(t,e){let n={};return Object.keys(e).forEach(a=>{const i=e[a];n[a]=t[i]}),n};var b=function(){const t=this.getResponseObj();if(t&&t.facets&&t.facets.multilevel){const e=t.facets.multilevel.bucket||[];return d(e)}return[]},y=function(t=[]){const e=this;let n={},a={};return"string"==typeof t&&(t=Array(t)),t.forEach((t,i)=>{t.split(" OR ").forEach((t,i)=>{const s=t.split(":");if(2===s.length){s[1]=s[1].replace(/(^")|("$)/g,"").replace(/\"{2,}/g,'"').replace(/\\\"/g,'"').replace(/(^\[)|(\]$)/g,"");if(s[1].indexOf("TO")>0){const t=s[1].split(" TO ");a[s[0]]="["+Number(t[0])+" TO "+Number(t[1])+"]"}else{const t=e.getFacetByValue(s[0],s[1]);if(t){const{facetName:e,selectedvalue:a}=t;n[e]?n[e].push(a):n[e]=Array(a)}}}})}),{selectedFilters:n,rangeFacet:a}},F=function(t){const e=t["facet.multilevel"];let n=t["category-filter"],a={};return e&&n&&(n=n.split(">"),a[e]=n),a},S=function(){const t=this.getResponseObj();if(t){const{facets:e}=t;return e&&e.range&&e.range.list?e.range.list:[]}return[]},E=function(){const t=this.getResponseObj();if(t){const{facets:e}=t;return e&&e.text&&e.text.list?e.text.list:[]}return[]},C=function(){const{selectedFacets:t}=this.state;return t},R=function(t){return this.getFacets().find(e=>{const{facetName:n}=e;return n===t?e:null})},I=function(t){const{selectedFacets:e}=this.state,{selectedFacetName:n}=t;let a="",i="";const s=this.getSelectedFacetValue(t);if(s){const{name:t,dataId:e}=s;a=t,i=e}e[n]?e[n].push({name:a,dataId:i}):e[n]=[{name:a,dataId:i}],this.setPageStart(0),this.getResults.bind(this)()},k=function(t){return this.state.selectedFacets[t]},w=function(t={}){let{parent:e=null,level:n=null,name:a=null}=t;n&&this.state.categoryFilter[e]&&(n=Number(n)-1,this.state.categoryFilter[e].splice(n,6))},T=function(t,e){const n=this.getSelectedFacets();if(e){const a=n[t];this.state.selectedFacets[t]=a.filter(t=>t.dataId!=e)}else delete this.state.selectedFacets[t];return this.state.selectedFacets},P=function(t,e){const n=this.getSelectedFacet(t);let a=null;const{values:i}=n,s=i.find((function(t){return t.dataId==e}));if(s){const e=this.findSelectedFacet(t);if(e){const{name:t}=s;a=e.find(e=>e===t)}}return a},L=function(t,e){const n=this.getSelectedFacet(t)||{facetName:t};let a=!1,i=null;const{values:s=[]}=n;let r={name:e};return s.length>0&&(a=!0,r=s.find((function(t){return t.name==e}))),r&&(i={...n,selectedvalue:r,actualFound:a}),i},O=function(t){const{selectedFacetId:e,facetData:n}=t;return n.values.find(t=>{const{dataId:n}=t;return n==e})},N=function(t={}){let{parent:e=null,level:n=null,name:a=null}=t;n&&(this.state.categoryFilter[e]?(n=Number(n),this.state.categoryFilter[e][n-1]=a):this.state.categoryFilter[e]=[a])},x=function(t){const{start:e,end:n,facetName:a}=t;this.state.rangeFacet[a]="["+e+" TO "+n+"]"},B=function(t){let e=this.state.rangeFacet;e[t]&&delete e[t],this.state.rangeFacet=e},j=function(t,e){this.deleteFacet(t,e),this.setPageStart(0),this.getResults.bind(this)()};const M=function(){return this.state.categoryFilter},A=function(){this.getResults.bind(this)()};var U=function(){let t=[],e=this.getBreadCrumbs();for(let n=0;n<this.options.facetDepth;n++){if(e.level){const{level:n,values:a,filterField:i}=e;t.push({level:n,filterField:i,value:a[0].value})}if(!e.child)break;e=e.child}return t},V=function(){const t=this.getResponseObj();return t&&t.facets&&t.facets.multilevel&&t.facets.multilevel.breadcrumb||[]};var _=function(t){this.setSort(t),this.setPageStart(0),this.getResults.bind(this)()},W=function(t){if(void 0!==t)this.state.selectedSort=t;else{const t=this.getSearchQueryParams();this.state.selectedSort=t.sort?t.sort:""}};const D=function(){return this.state.selectedSort};var H=function(){const t=this.getResponseObj();return t&&t.didYouMean?t.didYouMean:null},X=function(){const t=this.state.didYouMean;return t&&t[0]?t[0].suggestion:""};var q=function(){const t=this.getResponseObj();return t&&t.banner&&t.banner.banners&&t.banner.banners||[]};const Y=function(t,e){this.state.userInput=t,this.callBack(t,e)};var G=t=>{const{prototype:e}=t;e.changeInput=Y,(t=>{t=Object.assign(t,i)})(e),(t=>{Object.assign(t,{setPageStart:s,getPaginationInfo:r})})(e),(t=>{t=Object.assign(t,{getSearchResults:o,getSearchMeta:c,getSearchQueryParams:l,getSearchQuery:u,getResults:p,processVariantMap:f,getProductByPropValue:g,setStateFromData:m})})(e),(t=>{Object.assign(t,{getSwatches:v})})(e),(t=>{t=Object.assign(t,{getBucketedFacets:b,getSelectedBucketedFacet:M,getFilterFromParams:y,getCategoryFilterFromParams:F,getRangeFacets:S,getFacets:E,getSelectedFacets:C,getSelectedFacet:R,updateFacets:I,findSelectedFacet:k,deleteCategoryFilter:w,deleteFacet:T,getTheFacetTypeByNameId:P,getFacetByValue:L,getSelectedFacetValue:O,setCategoryFilter:N,setRangeFacet:x,clearARangeFacet:B,deleteAFacet:j,applyRangeFacet:A})})(e),(t=>{t=Object.assign(t,{getBreadCrumbsList:U,getBreadCrumbs:V})})(e),(t=>{t=Object.assign(t,{applySort:_,getSelectedSort:D,setSort:W})})(e),(t=>{t=Object.assign(t,{getDidYouMeanFromResponse:H,getSpellCheckSuggested:X})})(e),(t=>{t=Object.assign(t,{getBanners:q})})(e)};var Q=function t(e){if(!e)return e;let n,a=Array.isArray(e)?[]:{};for(const i in e)n=e[i],a[i]="object"==typeof n?t(n):n;return a};class z{constructor(t){this.options=Object.assign({},a,t),this.state={userInput:"",responseObj:null,selectedFacets:{},currentUrl:"",didYouMean:null,startPageNo:this.options.startPageNo||0,isLoading:!1,selectedSort:"",isUrlUpdated:!1,rangeSliderList:[],rangeFacet:{},categoryFilter:{},breadcrumbs:{}};const{searchQueryParam:e,productType:n}=this.options;let i="/search?";"CATEGORY"===n&&(i="/category?"),this.url=this.getBaseUrl()+i+e+"="}getResponseObj(){const{responseObj:t}=this.state;return t?Q(t):null}setUrl(){const t=this.options.searchQueryParam+this.state.currentUrl.split("?q")[1];location.hash=t}callBack(t,e){const{callBackFn:n}=this.options;n(this,e)}}G(z);var K=z;var J=function(t,e,n,a){t.addEventListener(e,(function(t){for(var e=t.target;e&&e!==this;)e.matches(n)&&a.call(e,t),e=e.parentNode}))};function $(t,e){for(var n=0;n<e.length;n++){var a=e[n];a.enumerable=a.enumerable||!1,a.configurable=!0,"value"in a&&(a.writable=!0),Object.defineProperty(t,a.key,a)}}var Z,tt={facetName:"v_price",displayName:"Price",position:20,gap:50,start:0,end:100,normalizeFact:0,expo:-2,minX:0,minY:100,showValues:!0},et=function(){function t(e){!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t),this.options=Object.assign({},tt,e),this.state={step:0,startX:0,startY:100,x:this.options.minX,y:this.options.minY,selectedTouch:null,initialLeft:0},this.id=this.options.facetName+"_rangeSlider",this.slideLeftId="slider-touch-left",this.slideRightId="slider-touch-right",this.slideLineId="slider-line",this.innerLineId="slider-inner-line",this.valueId="valueContainer",this.delegate(this.options.wrapper,"mousedown","."+this.slideLeftId,this.onStart.bind(this)),this.delegate(this.options.wrapper,"mousedown","."+this.slideRightId,this.onStart.bind(this)),this.delegate(this.options.wrapper,"touchstart","."+this.slideLeftId,this.onStart.bind(this)),this.delegate(this.options.wrapper,"touchstart","."+this.slideRightId,this.onStart.bind(this)),"function"==typeof this.options.onChange?this.options.onChange.bind(this):this.options.onChange=function(){},"function"==typeof this.options.onStop?this.options.onStop.bind(this):this.options.onStop=function(){},this.onChange=this.onChange.bind(this),this.onMove=this.onMove.bind(this),this.onStop=this.onStop.bind(this),this.onStart=this.onStart.bind(this),this.handleDocElemEvts.bind(this)}var e,n,a;return e=t,(n=[{key:"getValue",value:function(){var t=this.getRangeValue();return{left:t.x,right:t.y,props:this.options,max:this.maxX}}},{key:"onChange",value:function(t){var e=this.getValue.bind(this)();"stop"===t?this.options.onStop(e):this.options.onChange(e),this.options.showValues&&(this.valueUI.innerHTML=this.renderValuesUI())}},{key:"onMove",value:function(t){this.getActionType(t);var e=t;t.touches&&(e=t.touches[0]);var n=this.options.normalizeFact;this.state.x=e.pageX-this.state.startX,this.state.y=e.pageX-this.state.startY,this.state.selectedTouch===this.touchLeft&&(this.state.x>this.touchRight.offsetLeft-this.state.selectedTouch.offsetWidth+n?this.state.x=this.touchRight.offsetLeft-this.state.selectedTouch.offsetWidth+n:this.state.x<0&&(this.state.x=0),this.setPositionLeft.bind(this)()),this.state.selectedTouch===this.touchRight&&(this.state.y<this.touchLeft.offsetLeft+this.touchLeft.offsetWidth-n?this.state.y=this.touchLeft.offsetLeft+this.touchLeft.offsetWidth-n:this.state.y>this.maxX&&(this.state.y=this.maxX),this.setPositionRight.bind(this)()),this.setInnerLine(),this.valueUI&&(this.valueUI.innerHTML=this.renderValuesUI()),this.onChange()}},{key:"setPositionLeft",value:function(){this.touchLeft.style.left=this.state.x+"px"}},{key:"setPositionRight",value:function(){this.touchRight.style.left=this.state.y+"px"}},{key:"setInnerLine",value:function(){this.innerLine.style.marginLeft=this.touchLeft.offsetLeft+"px",this.innerLine.style.width=this.touchRight.offsetLeft-this.touchLeft.offsetLeft+"px"}},{key:"onStop",value:function(t){this.handleDocElemEvts(!1),this.setInnerLine(),this.state.selectedTouch=null,this.onChange("stop")}},{key:"getActionType",value:function(t){if(t&&t.target&&t.target.dataset)return t.target.dataset.action}},{key:"onStart",value:function(t){t.preventDefault();var e=t,n=this.getActionType(t);n&&("handleLeft"===n&&(this.state.x=this.touchLeft.offsetLeft,this.state.selectedTouch=this.touchLeft),"handleRight"===n&&(this.state.y=this.touchRight.offsetLeft,this.state.selectedTouch=this.touchRight)),this.state.startX=e.pageX-this.state.x,this.state.startY=e.pageX-this.state.y,this.handleDocElemEvts(!0)}},{key:"handleDocElemEvts",value:function(t){t?(this.wrapper.addEventListener("mousemove",this.onMove),this.wrapper.addEventListener("mouseup",this.onStop),this.wrapper.addEventListener("touchmove",this.onMove),this.wrapper.addEventListener("touchend",this.onStop)):(this.wrapper.removeEventListener("mousemove",this.onMove),this.wrapper.removeEventListener("mouseup",this.onStop),this.wrapper.removeEventListener("touchmove",this.onMove),this.wrapper.removeEventListener("touchend",this.onStop))}},{key:"delegate",value:function(t,e,n,a){t.addEventListener(e,(function(t){for(var e=t.target;e&&e!==this;)e.matches(n)&&a.call(e,t),e=e.parentNode}))}},{key:"setIntialPosition",value:function(){var t=this.options.end/100,e=this.maxX/100,n=this.state.x/t*e,a=this.state.y/t*e;this.state.x=Math.round(n),this.state.y=Math.round(a),this.innerLine.style.width=this.state.y-this.state.x+"px",this.state.initialLeft=this.state.x+this.options.normalizeFact,this.innerLine.style.marginLeft=this.state.initialLeft+"px",this.setPositionLeft(),this.setPositionRight(),this.options.showValues&&(this.valueUI.innerHTML=this.renderValuesUI())}},{key:"bindElements",value:function(){this.touchLeft||(this.touchLeft=document.querySelector("#"+this.id+" ."+this.slideLeftId)),this.touchRight||(this.touchRight=document.querySelector("#"+this.id+" ."+this.slideRightId)),this.maxX=this.options.wrapper.offsetWidth-(this.touchRight.offsetWidth+10),this.slideLine||(this.slideLine=document.querySelector("#"+this.id+" ."+this.slideLineId),this.slideLine.style.width=this.maxX-this.options.normalizeFact+"px"),this.innerLine||(this.innerLine=document.querySelector("#"+this.id+" ."+this.innerLineId)),this.wrapper||(this.wrapper=document.getElementById(this.id)),this.options.showValues&&!this.valueUI&&(this.valueUI=document.querySelector("#"+this.id+" ."+this.valueId)),this.setIntialPosition()}},{key:"getRangeValue",value:function(){var t=Number(this.touchLeft.style.left.replace("px","")),e=Number(this.touchRight.style.left.replace("px","")),n=this.options.end/this.maxX;return{x:Math.round(t*n),y:Math.round(e*n)}}},{key:"renderValuesUI",value:function(){var t=this.getRangeValue(),e=t.x,n=t.y;return'<div  style="display:flex">\n                <div>min : '.concat(e,"</div>\n                <div>msx : ").concat(n,"</div>\n            </div>")}},{key:"render",value:function(){var t=this.options,e=t.facetName,n=t.displayName;return setTimeout(this.bindElements.bind(this),10),'<div id="'.concat(this.id,'" class="range-slider-container">\n                ').concat(n,'\n                <div class="range-slider" >\n                    <div data-action="handleLeft" class="').concat(this.slideLeftId,' round-handle">\n                    </div>\n                    <div data-action="handleRight" class="').concat(this.slideRightId,' round-handle">\n                    </div>\n                    <div class="').concat(this.slideLineId,'">\n                        <div class="').concat(this.innerLineId,'"></div>\n                    </div>\n                </div>\n            <div>\n            <div class="valueContainer" >\n                ').concat("",'\n            </div>\n            <div>\n                <button data-facet-name="').concat(e,'" data-action="filterPriceRange"> filter by price </button>\n                <button data-facet-name="').concat(e,'"data-action="clearPriceRange"> clear </button>\n            <div>\n        </div>')}}])&&$(e.prototype,n),a&&$(e,a),t}();function nt(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var at=(nt(Z={beforeApiCall:"BEFORE_API_CALL",afterApiCall:"AFTER_API_CALL",beforeRender:"BEFORE_RENDER",beforeNoResultRender:"BEFORE_NO_RESULTS_RENDER",afterNoResultRender:"AFTER_NO_RESULTS_RENDER",afterRender:"AFTER_RENDER",deleteFacet:"DELETE_FACET",changeFacet:"CHANGE_FACET",facetClick:"FACETS_CLICK",deleteFacetValue:"DELETE_FACET_VALUE"},"deleteFacet","DELETE_FACET"),nt(Z,"clearSort","CLEAR_SORT"),nt(Z,"changeSort","CHANGE_SORT"),nt(Z,"pageNext","PAGE_NEXT"),nt(Z,"pagePrev","PAGE_PREV"),nt(Z,"changeInput","CHANGE_INPUT"),Z),it={clearCategoryFilter:"clearCategoryFilter",setCategoryFilter:"setCategoryFilter",changeSwatch:"changeSwatch",clearPriceRange:"clearPriceRange",filterPriceRange:"filterPriceRange",prev:"prev",next:"next",clearSort:"clearSort",changeSort:"changeSort",deleteFacet:"deleteFacet",deleteFacetValue:"deleteFacetValue"};function st(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var a=Object.getOwnPropertySymbols(t);e&&(a=a.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,a)}return n}function rt(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?st(Object(n),!0).forEach((function(e){ct(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):st(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function ct(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var ot,lt=function(t,e){var n=this,a="";return t.forEach((function(t){var e=t.level,i=t.filterField,s=t.value,r="".concat(n.bucketedFacetElem,"  category-level-").concat(e);a+='<button \n        data-parent="'.concat(i,'"\n        data-level="').concat(e,'"\n        data-name="').concat(s,'"\n        class=" ').concat(r,' selected-crumb"\n        data-action = "clearCategoryFilter">\n        ').concat(s," x</button>")})),a},ut=function(t){return'<div class=\'product-view-container\'>\n        <button id="listBtn" class="'.concat("LIST"===t?"selected-view":"",' view-btn" data-view-action="LIST" >List</button>\n        <button id="gridBtn" class="').concat("GRID"===t?"selected-view":"",' view-btn"  data-view-action="GRID" >Grid</button>\n    </div>')},ht=function(t){var e=t.map((function(t){var e=t.imageUrl;return'<div><img style="max-width:100%" src="'.concat(e,'"/></div>')})).join("");return"".concat(e)};function dt(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var pt=(dt(ot={productId:"uniqueId",searchBoxSelector:null,siteKey:"prod-kookai-com-au4541568774205",apiKey:"1c406c7058fdd75c04293fa87c2f9720",sdkHostName:"https://search.unbxd.io/",searchResultsTemplate:function(t){var e=t.title,n=t.sku;return'<div id="'.concat(n,'" class="product-item" style="border:solid 1px green">\n         ').concat(e,"\n        </div>")},searchResultsSelector:null,productItemClass:".product-item",facetWrapper:null,selectedFacetElem:function(t,e){var n=t.facetName,a=e.name,i=e.count,s=e.dataId;return'<option \n                selected\n                data-facet-name="'.concat(n,'"\n                data-facet-action="deleteFacetValue"\n                data-id= "').concat(s,'">\n                ').concat(a," (").concat(i,")\n        </option>")},facetElem:function(t,e){var n=t.facetName,a=e.name,i=e.count,s=e.dataId;return'<option\n                data-facet-name="'.concat(n,'" \n                data-facet-action="CHANGE_FACET"\n                class="').concat(this.selectedFacetClass,'"\n                data-id= "').concat(s,'">\n                    ').concat(a," (").concat(i,")\n            </option>")},facetItemElem:function(t,e){var n=t.displayName,a=t.facetName;return'<select multiple id="'.concat(a,'">\n                <option \n                    data-facet-action="deleteFacet"\n                    data-facet-name="').concat(a,'" >\n                        ').concat(n,"\n                        </option>\n                ").concat(e,"\n            </select>")},facetElemWrapClass:"select-facets-block",facetEvt:"change",selectedFacetBlock:null,selectedFacetClass:"selected-facet",productType:"SEARCH",searchQueryParam:"q",defaultFilters:null,spellCheckSelector:null,spellCheckTemplate:function(t){return"<div>\n        did you mean <strong>".concat(t,"</strong>\n    </div>")},noResultsUi:function(t){return"<div> No Results found ".concat(t," </div>")},noResultContainer:null,callBackFn:function(t,e){console.log(t,e,"state,type")},pageSize:10,startPageNo:0,paginationContainer:null,paginationUI:function(t){var e=t.noOfPages,n=t.currentPage,a='<button class="next-btn" data-page-action="next">next</button>',i='<button class="prev-btn" data-page-action="prev">prev</button>';return t.isNext||(a='<button disabled class="next-btn">next</button>'),t.isPrev||(i='<button disabled class="prev-btn">prev</button>'),'<div class="pagination-block">\n       '.concat(n," of ").concat(e," ----  ").concat(i," ").concat(a," \n    </div>")},paginationEvt:"click",sortContainer:null,sortOptions:[{value:"price desc",text:"Price High to Low"},{value:"price asc",text:" Price Low to High"},{value:"rating asc",text:" Rating Low to High"},{value:"rating desc",text:" Rating High to low"}],sortUI:function(t){var e="";return this.options.sortOptions.forEach((function(n){var a=n.value,i=n.text;e+=a==t?'<option value="'.concat(a,'" selected>\n                ').concat(i,"\n            </option>"):'<option value="'.concat(a,'">').concat(i,"</option>")})),'<div>\n        <select id="unbxdSorter" class="select-class">\n            <option value="">\n                Sort By Relevancy\n            </option>\n            '.concat(e,'\n        </select>\n        <button data-action="clearSort">clear sort</button>\n    </div>')},sortEvt:"change",sortElem:"select",productClick:function(t){console.log(t,"product,index")},fields:["title","price","uniqueId","sku","rating"],spellCheck:!1,facetMultiSelect:!1,facetMultiSelectionMode:!1,loaderElem:function(){return"<div>Loading search results....</div>"},loaderContainer:null,variants:!1,variantMapping:{},rangeFacetContainer:null,rangeFacetUI:function(t){var e=this,n=this,a=function(t){var n=t.left,a=t.right,i=t.props,s=((i=void 0===i?{}:i).end,{start:n,end:a,facetName:i.facetName,displayName:i.displayName,gap:i.gap});e.setRangeFilter(s)},i=this.state.rangeFacet,s=t.map((function(t){var e=t.facetName,s=t.displayName,r=t.position,c=t.values,o=i[e],l=c.start,u=c.end;if(o){var h=function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:0,n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:100,a=t.replace(/(^")|("$)/g,"").replace(/\"{2,}/g,'"').replace(/\\\"/g,'"').replace(/(^\[)|(\]$)/g,""),i=a.split(" TO "),s=e,r=n;return 2===i.length&&(s=Number(i[0]),r=Number(i[1])),{start:s,end:r}}(o,c.start,c.end);l=h.start,u=h.end}return new n.widgets.RangeSlider(rt(rt({facetName:e,displayName:s,position:r},c),{},{onStop:a,minX:l,minY:u,wrapper:n.options.rangeFacetContainer})).render()})).join("");return'<div class="range-facet">\n        '.concat(s,"\n    </div>")},extraParams:{version:"V2","facet.multilevel":"categoryPath","f.categoryPath.displayName":"category","f.categoryPath.max.depth":"4","f.categoryPath.facet.limit":"100"},facetMultilevel:!0,bucketedSearchUi:function(t,e,n){var a=this,i="",s="",r=t;return n&&(s=lt.bind(this)(n,e),i+=s),r.forEach((function(t){var e=t.level,n=t.multiLevelField,s=(t.displayName,t.values),r="category-level-".concat(e),c=s.map((function(t){var i=t.name,s=t.count;return'<button \n                data-parent="'.concat(n,'"\n                data-level="').concat(e,'"\n                class="').concat(a.bucketedFacetElem," ").concat(r,'"\n                data-name="').concat(i,'"\n                data-action = "setCategoryFilter">\n                ').concat(i,"-- ").concat(s,"</button>")}));i+="".concat(c.join(""))})),""!==i?'<div class="bucketed-facet-wrap">\n            <h6>Bucketed facets</h6>\n                '.concat(i,"\n        </div>"):""},bucketedFacetContainer:null,bucketFacetElem:"",bucketFacetEvnt:"click",bucketedFacetElem:"bucketFacetElem",facetDepth:2,breadcrumbContainer:null,breadcrump:!0,breadcrumbSelectorClass:"bread-crumb",breadCrumpTemplate:function(t){var e=this,n="";return t.forEach((function(t,a){var i=t.level,s=t.filterField,r=t.value,c="".concat(e.options.breadcrumbSelectorClass," bread-crumb-item");a>0&&(n+="<span> > </span>"),n+='<button \n        data-parent="'.concat(s,'"\n        data-level="').concat(i,'"\n        class="').concat(c,'"\n        data-name="').concat(r,'"\n        data-action = "clearCategoryFilter">\n        ').concat(r," x</button>")})),'<div class="bread-crumb-main">'.concat(n,"</div>")},swatches:!0,swatchMap:{},swatchTemplate:function(t){return"<div>swatchtemplate</div>"},paginationType:"FIXED_PAGINATION",inifinteScrollTriggerElem:window},"paginationContainer",document.getElementById("paginationContainer")),dt(ot,"productViewTypes","GRID"),dt(ot,"gridCount",3),dt(ot,"productViewTypeTemplate",ut),dt(ot,"productViewTypeSelector",null),dt(ot,"pageSizeOptions",[8,12,16,20,24]),dt(ot,"pageSizeDisplayAction","change"),dt(ot,"pageSizeTemplate",(function(t){return""})),dt(ot,"pageSizeContainerSelector",null),dt(ot,"bannerSelector",null),dt(ot,"bannerTemplate",ht),ot),ft=function(t,e){var n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},a=document.createElement(t);e&&a.setAttribute("id",e);var i=Object.keys(n);return i.forEach((function(t){a.setAttribute(t,n[t])})),a},gt=function(){this.searchResultsWrapper=ft("DIV","searchResultsBlock",{class:"search-results-block"}),this.facetsWrapper=ft("DIV","facetsContainer",{class:"facets-results-block"}),this.rangeFacetsWrapper=ft("DIV","rangeFacetsContainer",{class:"range-facets-block"}),this.bucketedSearchWrapper=ft("DIV","bucketedSearchContainer",{class:"bucketed-facets-block"}),this.bannerWrapper=ft("DIV","bannerWrapper",{class:"banner-block"}),this.breadcrumbWrapper=ft("DIV","breadcrumbContainer",{class:"breadcrumbs-block"}),this.options.selectedFacetElem.bind(this),this.options.bucketedSearchUi.bind(this),this.options.facetElem.bind(this),this.options.facetItemElem.bind(this),this.options.facetWrapper.appendChild(this.facetsWrapper),this.options.breadcrump&&(this.options.breadCrumpTemplate=this.options.breadCrumpTemplate.bind(this)),this.options.bucketedFacetContainer?this.options.bucketedFacetContainer.appendChild(this.bucketedSearchWrapper):this.options.facetWrapper.appendChild(this.bucketedSearchWrapper),this.options.rangeFacetContainer?this.options.rangeFacetContainer.appendChild(this.rangeFacetsWrapper):this.options.facetWrapper.appendChild(this.rangeFacetsWrapper),this.options.breadcrumbContainer&&this.options.breadcrumbContainer.appendChild(this.breadcrumbWrapper),this.options.swatchTemplate&&this.options.swatchTemplate.bind(this),this.options.bannerSelector&&this.options.bannerSelector.appendChild(this.bannerWrapper),this.options.bannerTemplate=this.options.bannerTemplate.bind(this),this.options.searchResultsTemplate=this.options.searchResultsTemplate.bind(this),this.options.searchResultsSelector.appendChild(this.searchResultsWrapper),this.loaderContainer=this.options.loaderContainer||this.searchResultsWrapper};var mt=function(){var t=this.getBanners();this.bannerWrapper.innerHTML=this.options.bannerTemplate(t)},vt=function(t,e){var n=this,a=t||this.getFacets(),i=e||this.getSelectedFacets(),s="",r=this.options.selectedFacetBlock,c=a.map((function(t){t.displayName;var e=t.facetName,a=t.values,c=void 0===a?[]:a,o="",l=i[e];if(c.length>0){var u=c.map((function(e,a){var i=e.name,c=!1;return l&&(c=l.some((function(t){return JSON.stringify(t.name)===JSON.stringify(i)}))),c?r?void(s+=n.options.selectedFacetElem(t,e)):n.options.selectedFacetElem(t,e):n.options.facetElem(t,e)}));o=n.options.facetItemElem(t,u.join(""))}return r&&(r.innerHTML=s),'<div data-id="'.concat(e,'">\n            ').concat(o,"\n        </div>")})).join("");return'<div class="'.concat(this.options.facetElemWrapClass,'">\n        ').concat(c,"\n    </div>")},bt=function(t){var e=t.target.dataset,n=e.actionType,a=e.facetName;n===this.events.deleteFacet&&this.findSelectedFacet(a)&&(this.deleteAFacet.bind(this)(a),this.options.callBackFn(this,n,{facetName:a}))},yt=function(t){var e=t.target,n=("click"===this.options.facetEvt?e:e.options[e.selectedIndex]).dataset,a=n.facetName,i=n.facetAction,s=n.id;if(i===this.events.changeFacet){var r=this.getSelectedFacet(a),c={selectedFacetName:a,selectedFacetId:s,facetData:r};this.options.callBackFn(this,this.events.facetClick,{facetName:a,facetData:r}),this.updateFacets(c)}i===this.actions.deleteFacetValue&&this.findSelectedFacet(a)&&(this.deleteAFacet.bind(this)(a,s),this.options.callBackFn(this,i,{facetName:a})),i===this.actions.deleteFacet&&this.findSelectedFacet(a)&&(this.deleteAFacet.bind(this)(a),this.options.callBackFn(this,i,{facetName:a}))},Ft=function(t){var e=t.target.dataset,n=e.action,a=e.facetName;n===this.actions.filterPriceRange&&(this.applyRangeFacet(),this.options.callBackFn(this,n,{facetName:a})),n===this.actions.clearPriceRange&&a&&(this.clearARangeFacet(a),this.getResults.bind(this)(),this.options.callBackFn(this,n,{facetName:a}))},St=function(t){var e=t.target.dataset;e.action===this.actions.setCategoryFilter&&(this.setCategoryFilter(e),this.options.callBackFn(this,e.action,e),this.getResults()),e.action===this.actions.clearCategoryFilter&&(this.deleteCategoryFilter(e),this.options.callBackFn(this,e.action,e),this.getResults())};function Et(){var t=this.getRangeFacets();return this.options.rangeFacetUI.bind(this)(t)}function Ct(t){this.setRangeFacet(t)}function Rt(){var t=this.getBucketedFacets(),e=this.getSelectedBucketedFacet(),n=this.getBreadCrumbsList();return this.options.bucketedSearchUi(t,e,n)}var It=function(){var t=this.getSearchResults();if(!t)return"";var e=t.products,n=this,a=this.options.gridCount,i=this.viewState.productViewType,s="";return"GRID"===i?e.forEach((function(t,e){var i=e%a;0===i&&(s+='<div class="unbxd-row">'),s+=n.options.searchResultsTemplate(t),i===a-1&&(s+="</div>")})):s=e.map((function(t){return n.options.searchResultsTemplate(t)})).join(""),"<div>\n        ".concat(s,"\n    </div>")},kt=function(t){var e=t.path,n=this.options.productItemClass.replace(".",""),a=t.target.dataset,i=e.find((function(t){return t.className.indexOf(n)>=0}));if(a.action===actions.changeSwatch)return this.options.callBackFn(this,a.action),i.querySelector(a.swatchTarget).src=a.swatchImg,!1;var s=this.getProductByPropValue(this.options.productId,i.id);this.options.productClick(s)};var wt=function(t){return this.options.noResultsUi.bind(this)(t)};var Tt=function(){if(this.options.productViewTypeSelector&&this.options.productViewTypeTemplate){var t=this.viewState.productViewType;this.options.productViewTypeSelector.innerHTML=this.options.productViewTypeTemplate.bind(this)(t)}};function Pt(t){var e=t.target.dataset.viewAction;this.viewState.productViewType=e,this.setPageStart(0),this.getResults.bind(this)()}var Lt=function(){var t="";return"FIXED_PAGINATION"===this.options.paginationType&&(t=this.options.paginationUI(this.getPaginationInfo())),t};function Ot(t){var e=this.getPaginationInfo(),n=e.start,a=(e.productsLn,e.numberOfProducts,e.rows),i=e.isNext,s=e.isPrev;if(t===this.actions.next){var r=n+a;i&&(this.setPageStart(r),this.getResults(),this.options.callBackFn(this,events.pageNext,{value:r}))}if(t===this.actions.prev){var c=n-a;s&&(this.setPageStart(c),this.getResults(),this.options.callBackFn(this,events.pagePrev,{value:c}))}}function Nt(t){var e=t.target.dataset.pageAction;this.renderNewResults(e)}function xt(t){var e=t.target,n=e.dataset.action,a=e.value;n===this.actions.clearSort&&(this.applySort(""),this.options.callBackFn(this,n)),n===this.actions.changeSort&&(this.applySort(a),this.options.callBackFn(this,n,{sort:a}))}var Bt=function(){return this.options.sortUI.bind(this)(this.getSelectedSort())},jt=function(t){var e=this.options.searchBoxSelector.value;this.changeInput(e,this.events.changeInput),e?this.getResults():this.searchResultsWrapper.innerHTML=null},Mt=function(t){return this.options.spellCheckTemplate.bind(this)(t)};function At(t){console.log(t,"breadcrumbs")}function Ut(){var t=this.getBreadCrumbsList();return this.options.breadCrumpTemplate(t)}function Vt(t){var e=this.getSwatches(t,this.options.swatchMap);return this.options.swatchTemplate(e)}var _t=function(t){var e=t.prototype;e.renderBannerUI=mt,function(t){t=Object.assign(t,{setInputValue:jt})}(e),function(t){t=Object.assign(t,{onPageViewTypeClick:Pt,renderProductViewTypeUI:Tt})}(e),function(t){t=Object.assign(t,{renderFacets:vt,facetsClickFn:bt,findChangedFacet:yt,onClickRangeFacet:Ft,renderRangeFacets:Et,setRangeFilter:Ct,renderBucketedUI:Rt,onBucketedFacet:St})}(e),function(t){t=Object.assign(t,{renderSearch:It,onProductItemClick:kt,renderNoResults:wt})}(e),function(t){t=Object.assign(t,{renderPagination:Lt,renderNewResults:Ot,paginationAction:Nt})}(e),function(t){t=Object.assign(t,{sortAction:xt,renderSort:Bt})}(e),function(t){t=Object.assign(t,{renderDidYouMean:Mt})}(e),function(t){t=Object.assign(t,{onBreadCrumbClick:At,renderBreadCrumbs:Ut})}(e),function(t){t=Object.assign(t,{renderSwatchBtns:Vt})}(e)},Wt=function(){var t=this.options,e=t.callBackFn,n=t.sortContainer,a=t.noResultContainer,i=t.spellCheckSelector,s=t.paginationContainer,r=this.events,c=r.beforeRender,o=r.beforeNoResultRender,l=r.afterNoResultRender,u=r.afterRender;e(this,c),this.loaderContainer.innerHTML=null;var h=this.getSearchResults();if(this.options.searchBoxSelector.value=this.state.userInput,this.facetsWrapper.innerHTML=this.renderFacets(),this.rangeFacetsWrapper.innerHTML=this.renderRangeFacets(),"INFINITE_SCROLL"===this.options.paginationType&&this.viewState.isInfiniteStarted?(this.viewState.isInfiniteStarted=!1,this.searchResultsWrapper.innerHTML+=this.renderSearch()):this.searchResultsWrapper.innerHTML=this.renderSearch(),this.bucketedSearchWrapper.innerHTML=this.renderBucketedUI(),this.breadcrumbWrapper.innerHTML=this.renderBreadCrumbs(),n&&(n.innerHTML=this.renderSort()),h&&0===h.numberOfProducts){e(this,o);var d=this.getSearchQuery();a&&(a.innerHTML=this.renderNoResults(d)),e(this,l)}else a.innerHTML=null;var p=this.getSpellCheckSuggested();i&&p&&(i.innerHTML=this.renderDidYouMean(p)),s.innerHTML=this.renderPagination(),this.renderProductViewTypeUI(),this.renderBannerUI(),e(this,u)};var Dt=function(){var t,e,n;n=document,Math.max(n.body.scrollHeight,n.documentElement.scrollHeight,n.body.offsetHeight,n.documentElement.offsetHeight,n.body.clientHeight,n.documentElement.clientHeight)-20<=(t=0,e=0,"number"==typeof window.pageYOffset?(e=window.pageYOffset,t=window.pageXOffset):document.body&&(document.body.scrollLeft||document.body.scrollTop)?(e=document.body.scrollTop,t=document.body.scrollLeft):document.documentElement&&(document.documentElement.scrollLeft||document.documentElement.scrollTop)&&(e=document.documentElement.scrollTop,t=document.documentElement.scrollLeft),[t,e])[1]+window.innerHeight&&(this.viewState.isInfiniteStarted||(this.viewState.isInfiniteStarted=!0,this.renderNewResults("next")))},Ht=function(t,e){var n;return function(){var a=this,i=arguments;clearTimeout(n),n=setTimeout((function(){return t.apply(a,i)}),e)}};var Xt=function(){var t=this,e=this.options,n=e.paginationContainer,a=e.paginationEvt,i=e.sortContainer,s=e.searchButtonSelector,r=e.searchEvt,c=e.productItemClass,o=e.sortElem,l=e.sortEvt,u=e.facetEvt,h=e.facetElemWrapClass,d=e.selectedFacetBlock,p=e.selectedFacetClass,f=e.productViewTypeSelector;n.addEventListener(a,this.paginationAction.bind(this)),s.addEventListener(r,this.setInputValue.bind(this)),this.delegate(this.searchResultsWrapper,"click",c,this.onProductItemClick.bind(this)),this.delegate(i,l,o,this.sortAction.bind(this)),this.delegate(this.facetsWrapper,u,"."+h,this.findChangedFacet.bind(this)),d&&this.delegate(d,u,"."+p,this.findChangedFacet.bind(this)),this.rangeFacetsWrapper&&this.delegate(this.rangeFacetsWrapper,"click","button",this.onClickRangeFacet.bind(this)),this.bucketedSearchWrapper&&this.delegate(this.bucketedSearchWrapper,this.options.bucketFacetEvnt,"."+this.options.bucketedFacetElem,this.onBucketedFacet.bind(this)),this.breadcrumbWrapper&&this.delegate(this.breadcrumbWrapper,"click","."+this.options.breadcrumbSelectorClass,this.onBucketedFacet.bind(this)),f&&this.delegate(f,"click","button",this.onPageViewTypeClick.bind(this)),"INFINITE_SCROLL"===this.options.paginationType&&document.addEventListener("scroll",Ht((function(){Dt.bind(t)()}),1e3))},qt=function(){this.renderDidYouMean.bind(this),this.options.callBackFn.bind(this),this.options.productClick.bind(this),this.callBack=this.callBack.bind(this),this.renderSwatchBtns=this.renderSwatchBtns.bind(this),this.reRender=Wt.bind(this),this.renderProductViewTypeUI(),this.getQueryParams()[this.options.searchQueryParam]&&this.renderFromUrl(),this.events=at,this.actions=it,this.bindEvents=Xt.bind(this)()};function Yt(t){return(Yt="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function Gt(t,e){for(var n=0;n<e.length;n++){var a=e[n];a.enumerable=a.enumerable||!1,a.configurable=!0,"value"in a&&(a.writable=!0),Object.defineProperty(t,a.key,a)}}function Qt(t,e){return(Qt=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function zt(t){var e=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(t){return!1}}();return function(){var n,a=$t(t);if(e){var i=$t(this).constructor;n=Reflect.construct(a,arguments,i)}else n=a.apply(this,arguments);return Kt(this,n)}}function Kt(t,e){return!e||"object"!==Yt(e)&&"function"!=typeof e?Jt(t):e}function Jt(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function $t(t){return($t=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}var Zt=function(t){!function(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&Qt(t,e)}(s,t);var e,n,a,i=zt(s);function s(t){var e;return function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,s),(e=i.call(this,t)).options=Object.assign({},pt,t),e.viewState={productViewType:e.options.productViewTypes,isInfiniteStarted:!1},gt.bind(Jt(e))(),qt.bind(Jt(e))(),e.widgets={RangeSlider:et.bind(Jt(e))},e}return e=s,(n=[{key:"callBack",value:function(t,e){var n=this.options,a=n.callBackFn,i=n.loaderElem,s=at.beforeApiCall,r=at.afterApiCall;e===s&&(a(this,s),this.loaderContainer.innerHTML=i(this)),e===r&&(a(this,r),this.reRender())}},{key:"delegate",value:function(t,e,n,a){return J(t,e,n,a)}}])&&Gt(e.prototype,n),a&&Gt(e,a),s}(K);_t(Zt);e.default=Zt},function(t,e,n){},function(t,e,n){"use strict";n.r(e);var a=n(0);n(1);console.log(a.default,"UnbxdSearchComponent");var i=new a.default({searchBoxSelector:document.getElementById("unbxdInput"),searchEvt:"click",searchButtonSelector:document.getElementById("searchBtn"),searchResultsSelector:document.getElementById("searchResultsWrapper"),facetWrapper:document.getElementById("facetsWrapper"),selectedFacetBlock:document.getElementById("selectedFacetWrapper"),siteKey:"prod-kookai-com-au4541568774205",apiKey:"1c406c7058fdd75c04293fa87c2f9720",sdkHostName:"https://search.unbxd.io/",producType:"SEARCH",searchQueryParam:"q",updateUrls:!0,productId:"uniqueId",searchResultsTemplate:function(t){var e=t.title,n=t.imageUrl,a=t.uniqueId,i=t.price,s=this.options,r=(s.swatchMap,"");return s.swatches&&(r=this.renderSwatchBtns(t)),'<div id="'.concat(a,'" data-item="product" class="product-item" style="border:solid 1px red;display:flex">  \n            <img class="productImgBlock" style="width:100px" src="').concat(n,'"/>\n        <div>\n         <h3>').concat(e," </h3>\n         <div>\n            ").concat(r,"\n         </div>\n         <strong>").concat(i,"</strong>\n         </div>\n        </div>")},productItemClass:".product-item",mapping:{title:""},fields:["title","uniqueId","price","sku","imageUrl"],platform:"IO",callBackFn:function(t,e,n){console.log(t,e,n,"unbxdCallbackEcma")},selectedFacetElem:function(t,e){var n=t.facetName,a=e.name,i=e.count,s=e.dataId;return'<div><button \n                    class="selected-facet-btn '.concat(this.selectedFacetClass,'"\n                    data-facet-name="').concat(n,'"\n                    data-facet-action ="deleteFacetValue"\n                    data-id= "').concat(s,'">\n                    ').concat(a," (").concat(i,')\n            </button><button class="').concat(this.selectedFacetClass,'"   data-id= "').concat(s,'" data-facet-action="deleteFacetValue" data-facet-name="').concat(n,'" > x</button></div>')},facetElem:function(t,e){var n=t.facetName,a=e.name,i=e.count,s=e.dataId;return'<button\n                    data-facet-name="'.concat(n,'" \n                    data-facet-action="CHANGE_FACET"\n                    data-id= "').concat(s,'">\n                        ').concat(a," (").concat(i,")\n                </button>")},facetItemElem:function(t,e){var n=t.displayName,a=t.facetName;return'<div id="'.concat(a,'">\n                    <h3> ').concat(n,'</h3>\n                    <button data-facet-action="deleteFacet" data-facet-name="').concat(a,'" > clear</button>\n                    ').concat(e,"\n                </div>")},facetElemWrapClass:"select-facets-block",facetEvt:"click",selectedFacetClass:"selected-facet",facetMultiSelect:!0,facetMultiSelectionMode:!0,defaultFilters:null,spellCheck:!0,spellCheckTemplate:function(t){return"<p>did you mean <strong>".concat(t,"</strong></p>")},spellCheckSelector:document.getElementById("didYouMeanWrapper"),noResultContainer:document.getElementById("noResultWrapper"),pageSize:12,paginationType:"FIXED_PAGINATION",heightDiffToTriggerNextPage:100,sortContainer:document.getElementById("sortWrapper"),sortOptions:[{value:"price desc",text:"Price High to Low"},{value:"price asc",text:" Price Low to High"}],sortUI:function(t){var e="";return this.options.sortOptions.forEach((function(n){var a=n.value,i=n.text,s="";a==t&&(s="selcted-sort-btn"),e+='<button class="btn '.concat(s,'" value="').concat(a,'" data-action="changeSort"> ').concat(i,"</button>")})),"<div>\n            <h6>sort selections</h6>\n            ".concat(e,'\n            <button data-action="clearSort">clear sort</button>\n        </div>')},sortEvt:"click",sortElem:"button",loaderElem:function(){return"<div>Loading....</div>"},loaderContainer:document.getElementById("loaderContainer"),variants:!0,variantMapping:{image_url:"v_image_url"},rangeFacetContainer:document.getElementById("rangeFacetWrapper"),facetMultilevel:!0,facetMultilevelName:"category",bucketedFacetElem:"bucketFacetElem",extraParams:{version:"V2","facet.multilevel":"categoryPath","f.categoryPath.displayName":"category","f.categoryPath.max.depth":"4","f.categoryPath.facet.limit":"100"},bucketedFacetContainer:document.getElementById("bucketedFacetWrapper"),facetDepth:4,breadcrumb:!0,breadcrumbContainer:document.getElementById("breadcrumpContainer"),breadcrumbSelectorClass:"bread-crumb",swatches:!0,swatchMap:{swatchList:"colours",swatchImgs:"variant_metadata",swatchColors:"unbxd_parentcolours"},swatchSelector:"swatchBtn",swatchTemplate:function(t){var e=t.swatchColors,n=void 0===e?[]:e,a=t.swatchImgs,i=void 0===a?[]:a,s=this.swatchSelector,r="";return n.forEach((function(t,e){var n=i[e];if(n){var a=n.split("::")[1];r+='<button \n                        data-swatch-id="'.concat(t,'"\n                        data-swatch-img="').concat(a,'" \n                        data-action="changeSwatch"\n                        data-swatch-target = ".productImgBlock"\n                        class="').concat(s,' swatch-btm"\n                        style="border:solid 1px ').concat(t,'"\n                        >').concat(t,"</button>")}})),'<div class="swatchContainer">'.concat(r,"</div>")},productViewTypes:"GRID",gridCount:4,productViewTypeSelector:document.getElementById("productViewTypeContainer"),productViewTypeAction:"click"});console.log(i,"es6unbxd")}]);
