+++ public/js/es6SDK.js
@@ -1,0 +1,1 @@
+!function(t){var e={};function n(i){if(e[i])return e[i].exports;var s=e[i]={i,l:!1,exports:{}};return t[i].call(s.exports,s,s.exports,n),s.l=!0,s.exports}n.m=t,n.c=e,n.d=function(t,e,i){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:i})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var i=Object.create(null);if(n.r(i),Object.defineProperty(i,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var s in t)n.d(i,s,function(e){return t[e]}.bind(null,s));return i},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=0)}([function(t,e,n){"use strict";n.r(e);var i={siteKey:"api key",apiKey:"site key",isFlag:!0,sdkHostName:"https://search.unbxd.io/",productType:"SEARCH",searchQueryParam:"q",fields:["title"],defaultFilters:null,spellCheck:!1,pageSize:5,startPageNo:0,facetMultiSelect:!0,facetMultiSelectionMode:!0,searchQueryParam:null,updateUrls:!0,variants:!1,variantMapping:{},variantConfig:{count:1},extraParams:"",facetMultilevel:!0,facetDepth:6,productId:"uniqueId",swatches:!0,swatchMap:{},callBackFn:()=>{}};const s={extraParamUrlString:function(){const{extraParams:t}=this.options;let e="";return Object.keys(t).forEach(n=>{e+="&"+n+"="+t[n]}),e},renderFromUrl:function(){let t=!1;const e=this.getQueryParams();if("object"==typeof e){const{q:n,filter:i,sort:s,start:a}=e;t=!0,this.state.userInput!==n&&(this.state.userInput=n);const r=this.getFilterFromParams(i);this.state.rangeFacet=r.rangeFacet,this.state.selectedFacets=r.selectedFilters,this.state.selectedSort=s||"",this.state.startPageNo=a||0,this.state.categoryFilter=this.getCategoryFilterFromParams(e)}t&&this.getResults()},getQueryParams:function(){const t=decodeURI(location.hash.substr(1)).split("&");let e=[];t.forEach((t,n)=>{t.indexOf("=")>0?e.push(t):e[n-1]=e[n-1]+"&"+t});let n={};for(let t=0;t<e.length;t++){const i=e[t].split("=");void 0===n[i[0]]?n[i[0]]=i[1]:"string"==typeof n[i[0]]?(n[i[0]]=[n[i[0]]],n[i[0]].push(i[1])):n[i[0]].push(i[1])}return n},urlFlattenFacets:function(){let t="";const{selectedFacets:e}=this.state;return Object.keys(e).forEach(n=>{const i=e[n];t+="&filter=",i.forEach((e,i)=>{const{name:s}=e;t+=0===i?n+':"'+JSON.stringify(s).replace('"',""):" OR "+n+':"'+JSON.stringify(s).replace('"',"")})}),t},getDefaultFiltersStr:function(){let t="";const e=this.options.defaultFilters;if(null!==e){Object.keys(e).forEach(n=>{t+="&filter="+n+":"+JSON.stringify(e[n])})}return t},getFieldsStr:function(){return this.options.fields?"&fields="+this.options.fields.join(","):""},getNoOfResultsStr:function(){return"&rows="+this.options.pageSize},getPageStartStr:function(){return"&start="+this.state.startPageNo},getNewUrlState:function(){const t=this.state.userInput,e=this.urlFlattenFacets(),n=this.getFieldsStr()+this.getDefaultFiltersStr(),i=this.getSortUrlString(),s=this.getRangeFilterStr(),a=this.categoryFilterUrlStr();let r=this.options.facetMultiSelect?"&facet.multiselect=true":"";return this.state.currentUrl=this.url+t+a+r+e+n+i+s+this.getNoOfResultsStr()+this.getPageStartStr()+this.extraParamUrlString(),this.state.currentUrl},getRangeFilterStr:function(){const{rangeFacet:t}=this.state,e=Object.keys(t);let n="";return e.forEach(e=>{n+="&filter="+e+":"+t[e]}),n},categoryFilterUrlStr:function(){const{categoryFilter:t}=this.state;let e="";return Object.keys(t).forEach((n,i)=>{const s=t[n].map((t,e)=>{let n=encodeURIComponent(t);return e>0&&(n=">"+n),n});e+="&category-filter="+s.join("")}),e},getSortUrlString:function(){const t=this.getSelectedSort();return t?"&sort="+encodeURI(t):""},getBaseUrl:function(){const{sdkHostName:t,siteKey:e,apiKey:n}=this.options;return t+n+"/"+e}};var a=function(t){this.state.startPageNo=t},r=function(){const t=this.getSearchResults(),{numberOfProducts:e,start:n,products:i}=t,s=this.options.pageSize,a=Math.floor(e/s)+1;let r=Math.floor(n/s)+1;const o=!(n+s>=e||n>=e),c=!(n-s<0||n<=0);return{numberOfProducts:e,start:n,productsLn:i.length,rows:s,noOfPages:a,currentPage:r,isNext:o,isPrev:c}};var o=function(){return this.state.responseObj.searchMetaData||null},c=function(){const t=this.getResponseObj();if(t){const{response:e}=t;return e||null}return null},l=function(){const t=this.getSearchMeta();return t&&t.queryParams||null},h=function(){return this.getSearchQueryParams().q||""};var u={changeInput:"CHANGE_INPUT",fetchError:"FETCH_ERROR",afterApiCall:"AFTER_API_CALL",beforeApiCall:"BEFORE_API_CALL"};const d=t=>{let e=[];return(t||[]).forEach((t,n)=>{const{values:i}=t;t.values=((t=[])=>{let e=[];return t.forEach((n,i)=>{const s=n;let a="";i%2==1&&(a=t[i-1],e.push({name:a,count:s,dataId:i}))}),e})(i),e.push(t)}),e};var p=function(t=""){const{userInput:e}=this.state,n=""===t?e:t;this.state.userInput=n,this.state.currentUrl=this.getNewUrlState(),this.state.isLoading=!0,this.callBack(this,u.beforeApiCall);const i=fetch(this.state.currentUrl),s=this;i.then(t=>t.json()).then(t=>{if(this.state.isLoading=!1,t){this.options.variants&&(t=this.processVariantMap(t)),this.state.responseObj=t;const e=this.getDidYouMeanFromResponse();if(this.options.spellCheck&&e){this.state.didYouMean=e;if(this.getSpellCheckSuggested())return this.getResults(this.getSpellCheckSuggested()),!1}const n=t.facets;if(n){const e=this.getFacets(),i=d.bind(this)(e);t.facets={...n,text:{list:i}}}this.state.responseObj=t,this.setSort(),this.setStateFromData(),this.setUrl(),this.callBack(s,u.afterApiCall)}}).catch(t=>{console.log(this,t),this.state.isLoading=!1,this.callBack(s,u.fetchError)})},f=function(t={}){const e=this;if(t.response&&t.response.products){let n=t.response.products.map(t=>{const{variants:n=[],relevantDocument:i}=t;let s=t;if("variant"===i&&n.length>0){const t=n[0],i=e.options.variantMapping;Object.keys(i).forEach(e=>{const n=i[e];s[e]=t[n]})}return s});t.response.products=n}return t},g=function(t,e){const n=this.getSearchResults(),{products:i}=n;return i.find(n=>n[t]===e)},v=function(){const t=this.getSearchQueryParams(),{q:e,filter:n}=t;this.state.userInput=e;const i=this.getFilterFromParams(n);this.state.rangeFacet=i.rangeFacet,this.state.selectedFacets=i.selectedFilters,this.state.categoryFilter=this.getCategoryFilterFromParams(t),this.state.breadcrumbs=this.getBreadCrumbs()};var b=function(t,e){let n={};return Object.keys(e).forEach(i=>{const s=e[i];n[i]=t[s]}),n};var m=function(){const t=this.getResponseObj();if(t&&t.facets&&t.facets.multilevel){const e=t.facets.multilevel.bucket||[];return d(e)}return[]},y=function(t=[]){const e=this;let n={},i={};return"string"==typeof t&&(t=Array(t)),t.forEach((t,s)=>{t.split(" OR ").forEach((t,s)=>{const a=t.split(":");if(2===a.length){a[1]=a[1].replace(/(^")|("$)/g,"").replace(/\"{2,}/g,'"').replace(/\\\"/g,'"').replace(/(^\[)|(\]$)/g,"");if(a[1].indexOf("TO")>0){const t=a[1].split(" TO ");i[a[0]]="["+Number(t[0])+" TO "+Number(t[1])+"]"}else{const t=e.getFacetByValue(a[0],a[1]);if(t){const{facetName:e,selectedvalue:i}=t;n[e]?n[e].push(i):n[e]=Array(i)}}}})}),{selectedFilters:n,rangeFacet:i}},F=function(t){const e=t["facet.multilevel"];let n=t["category-filter"],i={};return e&&n&&(n=n.split(">"),i[e]=n),i},S=function(){const t=this.getResponseObj();if(t){const{facets:e}=t;return e&&e.range&&e.range.list?e.range.list:[]}return[]},R=function(){const t=this.getResponseObj();if(t){const{facets:e}=t;return e&&e.text&&e.text.list?e.text.list:[]}return[]},C=function(){const{selectedFacets:t}=this.state;return t},E=function(t){return this.getFacets().find(e=>{const{facetName:n}=e;return n===t?e:null})},k=function(t){const{selectedFacets:e}=this.state,{selectedFacetName:n}=t;let i="",s="";const a=this.getSelectedFacetValue(t);if(a){const{name:t,dataId:e}=a;i=t,s=e}e[n]?e[n].push({name:i,dataId:s}):e[n]=[{name:i,dataId:s}],this.setPageStart(0),this.getResults.bind(this)()},I=function(t){return this.state.selectedFacets[t]},L=function(t={}){let{parent:e=null,level:n=null,name:i=null}=t;n&&this.state.categoryFilter[e]&&(n=Number(n)-1,this.state.categoryFilter[e].splice(n,6))},T=function(t,e){const n=this.getSelectedFacets();if(e){const i=n[t];this.state.selectedFacets[t]=i.filter(t=>t.dataId!=e)}else delete this.state.selectedFacets[t];return this.state.selectedFacets},w=function(t,e){const n=this.getSelectedFacet(t);let i=null;const{values:s}=n,a=s.find((function(t){return t.dataId==e}));if(a){const e=this.findSelectedFacet(t);if(e){const{name:t}=a;i=e.find(e=>e===t)}}return i},P=function(t,e){const n=this.getSelectedFacet(t)||{facetName:t};let i=!1,s=null;const{values:a=[]}=n;let r={name:e};return a.length>0&&(i=!0,r=a.find((function(t){return t.name==e}))),r&&(s={...n,selectedvalue:r,actualFound:i}),s},O=function(t){const{selectedFacetId:e,facetData:n}=t;return n.values.find(t=>{const{dataId:n}=t;return n==e})},N=function(t={}){let{parent:e=null,level:n=null,name:i=null}=t;n&&(this.state.categoryFilter[e]?(n=Number(n),this.state.categoryFilter[e][n-1]=i):this.state.categoryFilter[e]=[i])},x=function(t){const{start:e,end:n,facetName:i}=t;this.state.rangeFacet[i]="["+e+" TO "+n+"]"},j=function(t){let e=this.state.rangeFacet;e[t]&&delete e[t],this.state.rangeFacet=e},B=function(t,e){this.deleteFacet(t,e),this.setPageStart(0),this.getResults.bind(this)()};const M=function(){return this.state.categoryFilter},A=function(){this.getResults.bind(this)()};var U=function(){let t=[],e=this.getBreadCrumbs();for(let n=0;n<this.options.facetDepth;n++){if(e.level){const{level:n,values:i,filterField:s}=e;t.push({level:n,filterField:s,value:i[0].value})}if(!e.child)break;e=e.child}return t},V=function(){const t=this.getResponseObj();return t&&t.facets&&t.facets.multilevel&&t.facets.multilevel.breadcrumb||[]};var _=function(t){this.setSort(t),this.setPageStart(0),this.getResults.bind(this)()},D=function(t){if(void 0!==t)this.state.selectedSort=t;else{const t=this.getSearchQueryParams();this.state.selectedSort=t.sort?t.sort:""}};const W=function(){return this.state.selectedSort};var H=function(){const t=this.getResponseObj();return t&&t.didYouMean?t.didYouMean:null},X=function(){const t=this.state.didYouMean;return t&&t[0]?t[0].suggestion:""};var Y=function(){const t=this.getResponseObj();return t&&t.banner&&t.banner.banners&&t.banner.banners||[]};const q=function(t,e){this.state.userInput=t,this.callBack(t,e)};var G=t=>{const{prototype:e}=t;e.changeInput=q,(t=>{t=Object.assign(t,s)})(e),(t=>{Object.assign(t,{setPageStart:a,getPaginationInfo:r})})(e),(t=>{t=Object.assign(t,{getSearchResults:c,getSearchMeta:o,getSearchQueryParams:l,getSearchQuery:h,getResults:p,processVariantMap:f,getProductByPropValue:g,setStateFromData:v})})(e),(t=>{Object.assign(t,{getSwatches:b})})(e),(t=>{t=Object.assign(t,{getBucketedFacets:m,getSelectedBucketedFacet:M,getFilterFromParams:y,getCategoryFilterFromParams:F,getRangeFacets:S,getFacets:R,getSelectedFacets:C,getSelectedFacet:E,updateFacets:k,findSelectedFacet:I,deleteCategoryFilter:L,deleteFacet:T,getTheFacetTypeByNameId:w,getFacetByValue:P,getSelectedFacetValue:O,setCategoryFilter:N,setRangeFacet:x,clearARangeFacet:j,deleteAFacet:B,applyRangeFacet:A})})(e),(t=>{t=Object.assign(t,{getBreadCrumbsList:U,getBreadCrumbs:V})})(e),(t=>{t=Object.assign(t,{applySort:_,getSelectedSort:W,setSort:D})})(e),(t=>{t=Object.assign(t,{getDidYouMeanFromResponse:H,getSpellCheckSuggested:X})})(e),(t=>{t=Object.assign(t,{getBanners:Y})})(e)};var Q=function t(e){if(!e)return e;let n,i=Array.isArray(e)?[]:{};for(const s in e)n=e[s],i[s]="object"==typeof n?t(n):n;return i};class z{constructor(t){this.options=Object.assign({},i,t),this.state={userInput:"",responseObj:null,selectedFacets:{},currentUrl:"",didYouMean:null,startPageNo:this.options.startPageNo||0,isLoading:!1,selectedSort:"",isUrlUpdated:!1,rangeSliderList:[],rangeFacet:{},categoryFilter:{},breadcrumbs:{}};const{searchQueryParam:e,productType:n}=this.options;let s="/search?";"CATEGORY"===n&&(s="/category?"),this.url=this.getBaseUrl()+s+e+"="}getResponseObj(){const{responseObj:t}=this.state;return t?Q(t):null}setUrl(){const t=this.options.searchQueryParam+this.state.currentUrl.split("?q")[1];location.hash=t}callBack(t,e){const{callBackFn:n}=this.options;n(this,e)}}G(z);var K=z;var J=function(t,e,n,i){t.addEventListener(e,(function(t){for(var e=t.target;e&&e!==this;)e.matches(n)&&i.call(e,t),e=e.parentNode}))};function $(t,e){for(var n=0;n<e.length;n++){var i=e[n];i.enumerable=i.enumerable||!1,i.configurable=!0,"value"in i&&(i.writable=!0),Object.defineProperty(t,i.key,i)}}var Z,tt={facetName:"v_price",displayName:"Price",position:20,gap:50,start:0,end:100,normalizeFact:0,expo:-2,minX:0,minY:100,showValues:!0},et=function(){function t(e){!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t),this.options=Object.assign({},tt,e),this.state={step:0,startX:0,startY:100,x:this.options.minX,y:this.options.minY,selectedTouch:null,initialLeft:0},this.id=this.options.facetName+"_rangeSlider",this.slideLeftId="slider-touch-left",this.slideRightId="slider-touch-right",this.slideLineId="slider-line",this.innerLineId="slider-inner-line",this.valueId="valueContainer",this.delegate(this.options.wrapper,"mousedown","."+this.slideLeftId,this.onStart.bind(this)),this.delegate(this.options.wrapper,"mousedown","."+this.slideRightId,this.onStart.bind(this)),this.delegate(this.options.wrapper,"touchstart","."+this.slideLeftId,this.onStart.bind(this)),this.delegate(this.options.wrapper,"touchstart","."+this.slideRightId,this.onStart.bind(this)),"function"==typeof this.options.onChange?this.options.onChange.bind(this):this.options.onChange=function(){},"function"==typeof this.options.onStop?this.options.onStop.bind(this):this.options.onStop=function(){},this.onChange=this.onChange.bind(this),this.onMove=this.onMove.bind(this),this.onStop=this.onStop.bind(this),this.onStart=this.onStart.bind(this),this.handleDocElemEvts.bind(this)}var e,n,i;return e=t,(n=[{key:"getValue",value:function(){var t=this.getRangeValue();return{left:t.x,right:t.y,props:this.options,max:this.maxX}}},{key:"onChange",value:function(t){var e=this.getValue.bind(this)();"stop"===t?this.options.onStop(e):this.options.onChange(e),this.options.showValues&&(this.valueUI.innerHTML=this.renderValuesUI())}},{key:"onMove",value:function(t){this.getActionType(t);var e=t;t.touches&&(e=t.touches[0]);var n=this.options.normalizeFact;this.state.x=e.pageX-this.state.startX,this.state.y=e.pageX-this.state.startY,this.state.selectedTouch===this.touchLeft&&(this.state.x>this.touchRight.offsetLeft-this.state.selectedTouch.offsetWidth+n?this.state.x=this.touchRight.offsetLeft-this.state.selectedTouch.offsetWidth+n:this.state.x<0&&(this.state.x=0),this.setPositionLeft.bind(this)()),this.state.selectedTouch===this.touchRight&&(this.state.y<this.touchLeft.offsetLeft+this.touchLeft.offsetWidth-n?this.state.y=this.touchLeft.offsetLeft+this.touchLeft.offsetWidth-n:this.state.y>this.maxX&&(this.state.y=this.maxX),this.setPositionRight.bind(this)()),this.setInnerLine(),this.valueUI&&(this.valueUI.innerHTML=this.renderValuesUI()),this.onChange()}},{key:"setPositionLeft",value:function(){this.touchLeft.style.left=this.state.x+"px"}},{key:"setPositionRight",value:function(){this.touchRight.style.left=this.state.y+"px"}},{key:"setInnerLine",value:function(){this.innerLine.style.marginLeft=this.touchLeft.offsetLeft+"px",this.innerLine.style.width=this.touchRight.offsetLeft-this.touchLeft.offsetLeft+"px"}},{key:"onStop",value:function(t){this.handleDocElemEvts(!1),this.setInnerLine(),this.state.selectedTouch=null,this.onChange("stop")}},{key:"getActionType",value:function(t){if(t&&t.target&&t.target.dataset)return t.target.dataset.action}},{key:"onStart",value:function(t){t.preventDefault();var e=t,n=this.getActionType(t);n&&("handleLeft"===n&&(this.state.x=this.touchLeft.offsetLeft,this.state.selectedTouch=this.touchLeft),"handleRight"===n&&(this.state.y=this.touchRight.offsetLeft,this.state.selectedTouch=this.touchRight)),this.state.startX=e.pageX-this.state.x,this.state.startY=e.pageX-this.state.y,this.handleDocElemEvts(!0)}},{key:"handleDocElemEvts",value:function(t){t?(this.wrapper.addEventListener("mousemove",this.onMove),this.wrapper.addEventListener("mouseup",this.onStop),this.wrapper.addEventListener("touchmove",this.onMove),this.wrapper.addEventListener("touchend",this.onStop)):(this.wrapper.removeEventListener("mousemove",this.onMove),this.wrapper.removeEventListener("mouseup",this.onStop),this.wrapper.removeEventListener("touchmove",this.onMove),this.wrapper.removeEventListener("touchend",this.onStop))}},{key:"delegate",value:function(t,e,n,i){t.addEventListener(e,(function(t){for(var e=t.target;e&&e!==this;)e.matches(n)&&i.call(e,t),e=e.parentNode}))}},{key:"setIntialPosition",value:function(){var t=this.options.end/100,e=this.maxX/100,n=this.state.x/t*e,i=this.state.y/t*e;this.state.x=Math.round(n),this.state.y=Math.round(i),this.innerLine.style.width=this.state.y-this.state.x+"px",this.state.initialLeft=this.state.x+this.options.normalizeFact,this.innerLine.style.marginLeft=this.state.initialLeft+"px",this.setPositionLeft(),this.setPositionRight(),this.options.showValues&&(this.valueUI.innerHTML=this.renderValuesUI())}},{key:"bindElements",value:function(){this.touchLeft||(this.touchLeft=document.querySelector("#"+this.id+" ."+this.slideLeftId)),this.touchRight||(this.touchRight=document.querySelector("#"+this.id+" ."+this.slideRightId)),this.maxX=this.options.wrapper.offsetWidth-(this.touchRight.offsetWidth+10),this.slideLine||(this.slideLine=document.querySelector("#"+this.id+" ."+this.slideLineId),this.slideLine.style.width=this.maxX-this.options.normalizeFact+"px"),this.innerLine||(this.innerLine=document.querySelector("#"+this.id+" ."+this.innerLineId)),this.wrapper||(this.wrapper=document.getElementById(this.id)),this.options.showValues&&!this.valueUI&&(this.valueUI=document.querySelector("#"+this.id+" ."+this.valueId)),this.setIntialPosition()}},{key:"getRangeValue",value:function(){var t=Number(this.touchLeft.style.left.replace("px","")),e=Number(this.touchRight.style.left.replace("px","")),n=this.options.end/this.maxX;return{x:Math.round(t*n),y:Math.round(e*n)}}},{key:"renderValuesUI",value:function(){var t=this.getRangeValue(),e=t.x,n=t.y;return'<div  style="display:flex">\n                <div>min : '.concat(e,"</div>\n                <div>msx : ").concat(n,"</div>\n            </div>")}},{key:"render",value:function(){var t=this.options,e=t.facetName,n=t.displayName;return setTimeout(this.bindElements.bind(this),10),'<div id="'.concat(this.id,'" class="range-slider-container">\n                ').concat(n,'\n                <div class="range-slider" >\n                    <div data-action="handleLeft" class="').concat(this.slideLeftId,' round-handle">\n                    </div>\n                    <div data-action="handleRight" class="').concat(this.slideRightId,' round-handle">\n                    </div>\n                    <div class="').concat(this.slideLineId,'">\n                        <div class="').concat(this.innerLineId,'"></div>\n                    </div>\n                </div>\n            <div>\n            <div class="valueContainer" >\n                ').concat("",'\n            </div>\n            <div>\n                <button data-facet-name="').concat(e,'" data-action="filterPriceRange"> filter by price </button>\n                <button data-facet-name="').concat(e,'"data-action="clearPriceRange"> clear </button>\n            <div>\n        </div>')}}])&&$(e.prototype,n),i&&$(e,i),t}();function nt(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var it=(nt(Z={beforeApiCall:"BEFORE_API_CALL",afterApiCall:"AFTER_API_CALL",beforeRender:"BEFORE_RENDER",beforeNoResultRender:"BEFORE_NO_RESULTS_RENDER",afterNoResultRender:"AFTER_NO_RESULTS_RENDER",afterRender:"AFTER_RENDER",deleteFacet:"DELETE_FACET",changeFacet:"CHANGE_FACET",facetClick:"FACETS_CLICK",deleteFacetValue:"DELETE_FACET_VALUE"},"deleteFacet","DELETE_FACET"),nt(Z,"clearSort","CLEAR_SORT"),nt(Z,"changeSort","CHANGE_SORT"),nt(Z,"pageNext","PAGE_NEXT"),nt(Z,"pagePrev","PAGE_PREV"),nt(Z,"changeInput","CHANGE_INPUT"),Z),st={clearCategoryFilter:"clearCategoryFilter",setCategoryFilter:"setCategoryFilter",changeSwatch:"changeSwatch",clearPriceRange:"clearPriceRange",filterPriceRange:"filterPriceRange",prev:"prev",next:"next",clearSort:"clearSort",changeSort:"changeSort",deleteFacet:"deleteFacet",deleteFacetValue:"deleteFacetValue"};function at(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var i=Object.getOwnPropertySymbols(t);e&&(i=i.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,i)}return n}function rt(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?at(Object(n),!0).forEach((function(e){ot(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):at(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function ot(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var ct,lt=function(t,e){var n=this,i="";return t.forEach((function(t){var e=t.level,s=t.filterField,a=t.value,r="".concat(n.bucketedFacetElem,"  category-level-").concat(e);i+='<button \n        data-parent="'.concat(s,'"\n        data-level="').concat(e,'"\n        data-name="').concat(a,'"\n        class=" ').concat(r,' selected-crumb"\n        data-action = "clearCategoryFilter">\n        ').concat(a," x</button>")})),i},ht=function(t){return'<div class=\'product-view-container\'>\n        <button id="listBtn" class="'.concat("LIST"===t?"selected-view":"",' view-btn" data-view-action="LIST" >List</button>\n        <button id="gridBtn" class="').concat("GRID"===t?"selected-view":"",' view-btn"  data-view-action="GRID" >Grid</button>\n    </div>')},ut=function(t){var e=t.map((function(t){var e=t.imageUrl;return'<div><img style="max-width:100%" src="'.concat(e,'"/></div>')})).join("");return"".concat(e)};function dt(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var pt=(dt(ct={productId:"uniqueId",searchBoxSelector:null,siteKey:"prod-kookai-com-au4541568774205",apiKey:"1c406c7058fdd75c04293fa87c2f9720",sdkHostName:"https://search.unbxd.io/",searchResultsTemplate:function(t){var e=t.title,n=t.sku;return'<div id="'.concat(n,'" class="product-item" style="border:solid 1px green">\n         ').concat(e,"\n        </div>")},searchResultsSelector:null,productItemClass:".product-item",facetWrapper:null,selectedFacetElem:function(t,e){var n=t.facetName,i=e.name,s=e.count,a=e.dataId;return'<option \n                selected\n                data-facet-name="'.concat(n,'"\n                data-facet-action="deleteFacetValue"\n                data-id= "').concat(a,'">\n                ').concat(i," (").concat(s,")\n        </option>")},facetElem:function(t,e){var n=t.facetName,i=e.name,s=e.count,a=e.dataId;return'<option\n                data-facet-name="'.concat(n,'" \n                data-facet-action="CHANGE_FACET"\n                class="').concat(this.selectedFacetClass,'"\n                data-id= "').concat(a,'">\n                    ').concat(i," (").concat(s,")\n            </option>")},facetItemElem:function(t,e){var n=t.displayName,i=t.facetName;return'<select multiple id="'.concat(i,'">\n                <option \n                    data-facet-action="deleteFacet"\n                    data-facet-name="').concat(i,'" >\n                        ').concat(n,"\n                        </option>\n                ").concat(e,"\n            </select>")},facetElemWrapClass:"select-facets-block",facetEvt:"change",selectedFacetBlock:null,selectedFacetClass:"selected-facet",productType:"SEARCH",searchQueryParam:"q",defaultFilters:null,spellCheckSelector:null,spellCheckTemplate:function(t){return"<div>\n        did you mean <strong>".concat(t,"</strong>\n    </div>")},noResultsUi:function(t){return"<div> No Results found ".concat(t," </div>")},noResultContainer:null,callBackFn:function(t,e){console.log(t,e,"state,type")},pageSize:10,startPageNo:0,paginationContainer:null,paginationUI:function(t){var e=t.noOfPages,n=t.currentPage,i='<button class="next-btn" data-page-action="next">next</button>',s='<button class="prev-btn" data-page-action="prev">prev</button>';return t.isNext||(i='<button disabled class="next-btn">next</button>'),t.isPrev||(s='<button disabled class="prev-btn">prev</button>'),'<div class="pagination-block">\n       '.concat(n," of ").concat(e," ----  ").concat(s," ").concat(i," \n    </div>")},paginationEvt:"click",sortContainer:null,sortOptions:[{value:"price desc",text:"Price High to Low"},{value:"price asc",text:" Price Low to High"},{value:"rating asc",text:" Rating Low to High"},{value:"rating desc",text:" Rating High to low"}],sortUI:function(t){var e="";return this.options.sortOptions.forEach((function(n){var i=n.value,s=n.text;e+=i==t?'<option value="'.concat(i,'" selected>\n                ').concat(s,"\n            </option>"):'<option value="'.concat(i,'">').concat(s,"</option>")})),'<div>\n        <select id="unbxdSorter" class="select-class">\n            <option value="">\n                Sort By Relevancy\n            </option>\n            '.concat(e,'\n        </select>\n        <button data-action="clearSort">clear sort</button>\n    </div>')},sortEvt:"change",sortElem:"select",productClick:function(t){console.log(t,"product,index")},fields:["title","price","uniqueId","sku","rating"],spellCheck:!1,facetMultiSelect:!1,facetMultiSelectionMode:!1,loaderElem:function(){return"<div>Loading search results....</div>"},loaderContainer:null,variants:!1,variantMapping:{},rangeFacetContainer:null,rangeFacetUI:function(t){var e=this,n=this,i=function(t){var n=t.left,i=t.right,s=t.props,a=((s=void 0===s?{}:s).end,{start:n,end:i,facetName:s.facetName,displayName:s.displayName,gap:s.gap});e.setRangeFilter(a)},s=this.state.rangeFacet,a=t.map((function(t){var e=t.facetName,a=t.displayName,r=t.position,o=t.values,c=s[e],l=o.start,h=o.end;if(c){var u=function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:0,n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:100,i=t.replace(/(^")|("$)/g,"").replace(/\"{2,}/g,'"').replace(/\\\"/g,'"').replace(/(^\[)|(\]$)/g,""),s=i.split(" TO "),a=e,r=n;return 2===s.length&&(a=Number(s[0]),r=Number(s[1])),{start:a,end:r}}(c,o.start,o.end);l=u.start,h=u.end}return new n.widgets.RangeSlider(rt(rt({facetName:e,displayName:a,position:r},o),{},{onStop:i,minX:l,minY:h,wrapper:n.options.rangeFacetContainer})).render()})).join("");return'<div class="range-facet">\n        '.concat(a,"\n    </div>")},extraParams:{version:"V2","facet.multilevel":"categoryPath","f.categoryPath.displayName":"category","f.categoryPath.max.depth":"4","f.categoryPath.facet.limit":"100"},facetMultilevel:!0,bucketedSearchUi:function(t,e,n){var i=this,s="",a="",r=t;return n&&(a=lt.bind(this)(n,e),s+=a),r.forEach((function(t){var e=t.level,n=t.multiLevelField,a=(t.displayName,t.values),r="category-level-".concat(e),o=a.map((function(t){var s=t.name,a=t.count;return'<button \n                data-parent="'.concat(n,'"\n                data-level="').concat(e,'"\n                class="').concat(i.bucketedFacetElem," ").concat(r,'"\n                data-name="').concat(s,'"\n                data-action = "setCategoryFilter">\n                ').concat(s,"-- ").concat(a,"</button>")}));s+="".concat(o.join(""))})),""!==s?'<div class="bucketed-facet-wrap">\n            <h6>Bucketed facets</h6>\n                '.concat(s,"\n        </div>"):""},bucketedFacetContainer:null,bucketFacetElem:"",bucketFacetEvnt:"click",bucketedFacetElem:"bucketFacetElem",facetDepth:2,breadcrumbContainer:null,breadcrump:!0,breadcrumbSelectorClass:"bread-crumb",breadCrumpTemplate:function(t){var e=this,n="";return t.forEach((function(t,i){var s=t.level,a=t.filterField,r=t.value,o="".concat(e.options.breadcrumbSelectorClass," bread-crumb-item");i>0&&(n+="<span> > </span>"),n+='<button \n        data-parent="'.concat(a,'"\n        data-level="').concat(s,'"\n        class="').concat(o,'"\n        data-name="').concat(r,'"\n        data-action = "clearCategoryFilter">\n        ').concat(r," x</button>")})),'<div class="bread-crumb-main">'.concat(n,"</div>")},swatches:!0,swatchMap:{},swatchTemplate:function(t){return"<div>swatchtemplate</div>"},paginationType:"FIXED_PAGINATION",inifinteScrollTriggerElem:window},"paginationContainer",document.getElementById("paginationContainer")),dt(ct,"productViewTypes","GRID"),dt(ct,"gridCount",3),dt(ct,"productViewTypeTemplate",ht),dt(ct,"productViewTypeSelector",null),dt(ct,"pageSizeOptions",[8,12,16,20,24]),dt(ct,"pageSizeDisplayAction","change"),dt(ct,"pageSizeTemplate",(function(t){return""})),dt(ct,"pageSizeContainerSelector",null),dt(ct,"bannerSelector",null),dt(ct,"bannerTemplate",ut),ct),ft=function(t,e){var n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},i=document.createElement(t);e&&i.setAttribute("id",e);var s=Object.keys(n);return s.forEach((function(t){i.setAttribute(t,n[t])})),i},gt=function(){this.searchResultsWrapper=ft("DIV","searchResultsBlock",{class:"search-results-block"}),this.facetsWrapper=ft("DIV","facetsContainer",{class:"facets-results-block"}),this.rangeFacetsWrapper=ft("DIV","rangeFacetsContainer",{class:"range-facets-block"}),this.bucketedSearchWrapper=ft("DIV","bucketedSearchContainer",{class:"bucketed-facets-block"}),this.bannerWrapper=ft("DIV","bannerWrapper",{class:"banner-block"}),this.breadcrumbWrapper=ft("DIV","breadcrumbContainer",{class:"breadcrumbs-block"}),this.options.selectedFacetElem.bind(this),this.options.bucketedSearchUi.bind(this),this.options.facetElem.bind(this),this.options.facetItemElem.bind(this),this.options.facetWrapper.appendChild(this.facetsWrapper),this.options.breadcrump&&(this.options.breadCrumpTemplate=this.options.breadCrumpTemplate.bind(this)),this.options.bucketedFacetContainer?this.options.bucketedFacetContainer.appendChild(this.bucketedSearchWrapper):this.options.facetWrapper.appendChild(this.bucketedSearchWrapper),this.options.rangeFacetContainer?this.options.rangeFacetContainer.appendChild(this.rangeFacetsWrapper):this.options.facetWrapper.appendChild(this.rangeFacetsWrapper),this.options.breadcrumbContainer&&this.options.breadcrumbContainer.appendChild(this.breadcrumbWrapper),this.options.swatchTemplate&&this.options.swatchTemplate.bind(this),this.options.bannerSelector&&this.options.bannerSelector.appendChild(this.bannerWrapper),this.options.bannerTemplate=this.options.bannerTemplate.bind(this),this.options.searchResultsTemplate=this.options.searchResultsTemplate.bind(this),this.options.searchResultsSelector.appendChild(this.searchResultsWrapper),this.loaderContainer=this.options.loaderContainer||this.searchResultsWrapper};var vt=function(){var t=this.getBanners();this.bannerWrapper.innerHTML=this.options.bannerTemplate(t)},bt=function(t,e){var n=this,i=t||this.getFacets(),s=e||this.getSelectedFacets(),a="",r=this.options.selectedFacetBlock,o=i.map((function(t){t.displayName;var e=t.facetName,i=t.values,o=void 0===i?[]:i,c="",l=s[e];if(o.length>0){var h=o.map((function(e,i){var s=e.name,o=!1;return l&&(o=l.some((function(t){return JSON.stringify(t.name)===JSON.stringify(s)}))),o?r?void(a+=n.options.selectedFacetElem(t,e)):n.options.selectedFacetElem(t,e):n.options.facetElem(t,e)}));c=n.options.facetItemElem(t,h.join(""))}return r&&(r.innerHTML=a),'<div data-id="'.concat(e,'">\n            ').concat(c,"\n        </div>")})).join("");return'<div class="'.concat(this.options.facetElemWrapClass,'">\n        ').concat(o,"\n    </div>")},mt=function(t){var e=t.target.dataset,n=e.actionType,i=e.facetName;n===this.events.deleteFacet&&this.findSelectedFacet(i)&&(this.deleteAFacet.bind(this)(i),this.options.callBackFn(this,n,{facetName:i}))},yt=function(t){var e=t.target,n=("click"===this.options.facetEvt?e:e.options[e.selectedIndex]).dataset,i=n.facetName,s=n.facetAction,a=n.id;if(s===this.events.changeFacet){var r=this.getSelectedFacet(i),o={selectedFacetName:i,selectedFacetId:a,facetData:r};this.options.callBackFn(this,this.events.facetClick,{facetName:i,facetData:r}),this.updateFacets(o)}s===this.actions.deleteFacetValue&&this.findSelectedFacet(i)&&(this.deleteAFacet.bind(this)(i,a),this.options.callBackFn(this,s,{facetName:i})),s===this.actions.deleteFacet&&this.findSelectedFacet(i)&&(this.deleteAFacet.bind(this)(i),this.options.callBackFn(this,s,{facetName:i}))},Ft=function(t){var e=t.target.dataset,n=e.action,i=e.facetName;n===this.actions.filterPriceRange&&(this.applyRangeFacet(),this.options.callBackFn(this,n,{facetName:i})),n===this.actions.clearPriceRange&&i&&(this.clearARangeFacet(i),this.getResults.bind(this)(),this.options.callBackFn(this,n,{facetName:i}))},St=function(t){var e=t.target.dataset;e.action===this.actions.setCategoryFilter&&(this.setCategoryFilter(e),this.options.callBackFn(this,e.action,e),this.getResults()),e.action===this.actions.clearCategoryFilter&&(this.deleteCategoryFilter(e),this.options.callBackFn(this,e.action,e),this.getResults())};function Rt(){var t=this.getRangeFacets();return this.options.rangeFacetUI.bind(this)(t)}function Ct(t){this.setRangeFacet(t)}function Et(){var t=this.getBucketedFacets(),e=this.getSelectedBucketedFacet(),n=this.getBreadCrumbsList();return this.options.bucketedSearchUi(t,e,n)}var kt=function(){var t=this.getSearchResults();if(!t)return"";var e=t.products,n=this,i=this.options.gridCount,s=this.viewState.productViewType,a="";return"GRID"===s?e.forEach((function(t,e){var s=e%i;0===s&&(a+='<div class="unbxd-row">'),a+=n.options.searchResultsTemplate(t),s===i-1&&(a+="</div>")})):a=e.map((function(t){return n.options.searchResultsTemplate(t)})).join(""),"<div>\n        ".concat(a,"\n    </div>")},It=function(t){var e=t.path,n=this.options.productItemClass.replace(".",""),i=t.target.dataset,s=e.find((function(t){return t.className.indexOf(n)>=0}));if(i.action===actions.changeSwatch)return this.options.callBackFn(this,i.action),s.querySelector(i.swatchTarget).src=i.swatchImg,!1;var a=this.getProductByPropValue(this.options.productId,s.id);this.options.productClick(a)};var Lt=function(t){return this.options.noResultsUi.bind(this)(t)};var Tt=function(){if(this.options.productViewTypeSelector&&this.options.productViewTypeTemplate){var t=this.viewState.productViewType;this.options.productViewTypeSelector.innerHTML=this.options.productViewTypeTemplate.bind(this)(t)}};function wt(t){var e=t.target.dataset.viewAction;this.viewState.productViewType=e,this.setPageStart(0),this.getResults.bind(this)()}var Pt=function(){var t="";return"FIXED_PAGINATION"===this.options.paginationType&&(t=this.options.paginationUI(this.getPaginationInfo())),t};function Ot(t){var e=this.getPaginationInfo(),n=e.start,i=(e.productsLn,e.numberOfProducts,e.rows),s=e.isNext,a=e.isPrev;if(t===this.actions.next){var r=n+i;s&&(this.setPageStart(r),this.getResults(),this.options.callBackFn(this,events.pageNext,{value:r}))}if(t===this.actions.prev){var o=n-i;a&&(this.setPageStart(o),this.getResults(),this.options.callBackFn(this,events.pagePrev,{value:o}))}}function Nt(t){var e=t.target.dataset.pageAction;this.renderNewResults(e)}function xt(t){var e=t.target,n=e.dataset.action,i=e.value;n===this.actions.clearSort&&(this.applySort(""),this.options.callBackFn(this,n)),n===this.actions.changeSort&&(this.applySort(i),this.options.callBackFn(this,n,{sort:i}))}var jt=function(){return this.options.sortUI.bind(this)(this.getSelectedSort())},Bt=function(t){var e=this.options.searchBoxSelector.value;this.changeInput(e,this.events.changeInput),e?this.getResults():this.searchResultsWrapper.innerHTML=null},Mt=function(t){return this.options.spellCheckTemplate.bind(this)(t)};function At(t){console.log(t,"breadcrumbs")}function Ut(){var t=this.getBreadCrumbsList();return this.options.breadCrumpTemplate(t)}function Vt(t){var e=this.getSwatches(t,this.options.swatchMap);return this.options.swatchTemplate(e)}var _t=function(t){var e=t.prototype;e.renderBannerUI=vt,function(t){t=Object.assign(t,{setInputValue:Bt})}(e),function(t){t=Object.assign(t,{onPageViewTypeClick:wt,renderProductViewTypeUI:Tt})}(e),function(t){t=Object.assign(t,{renderFacets:bt,facetsClickFn:mt,findChangedFacet:yt,onClickRangeFacet:Ft,renderRangeFacets:Rt,setRangeFilter:Ct,renderBucketedUI:Et,onBucketedFacet:St})}(e),function(t){t=Object.assign(t,{renderSearch:kt,onProductItemClick:It,renderNoResults:Lt})}(e),function(t){t=Object.assign(t,{renderPagination:Pt,renderNewResults:Ot,paginationAction:Nt})}(e),function(t){t=Object.assign(t,{sortAction:xt,renderSort:jt})}(e),function(t){t=Object.assign(t,{renderDidYouMean:Mt})}(e),function(t){t=Object.assign(t,{onBreadCrumbClick:At,renderBreadCrumbs:Ut})}(e),function(t){t=Object.assign(t,{renderSwatchBtns:Vt})}(e)},Dt=function(){var t=this.options,e=t.callBackFn,n=t.sortContainer,i=t.noResultContainer,s=t.spellCheckSelector,a=t.paginationContainer,r=this.events,o=r.beforeRender,c=r.beforeNoResultRender,l=r.afterNoResultRender,h=r.afterRender;e(this,o),this.loaderContainer.innerHTML=null;var u=this.getSearchResults();if(this.options.searchBoxSelector.value=this.state.userInput,this.facetsWrapper.innerHTML=this.renderFacets(),this.rangeFacetsWrapper.innerHTML=this.renderRangeFacets(),"INFINITE_SCROLL"===this.options.paginationType&&this.viewState.isInfiniteStarted?(this.viewState.isInfiniteStarted=!1,this.searchResultsWrapper.innerHTML+=this.renderSearch()):this.searchResultsWrapper.innerHTML=this.renderSearch(),this.bucketedSearchWrapper.innerHTML=this.renderBucketedUI(),this.breadcrumbWrapper.innerHTML=this.renderBreadCrumbs(),n&&(n.innerHTML=this.renderSort()),u&&0===u.numberOfProducts){e(this,c);var d=this.getSearchQuery();i&&(i.innerHTML=this.renderNoResults(d)),e(this,l)}else i.innerHTML=null;var p=this.getSpellCheckSuggested();s&&p&&(s.innerHTML=this.renderDidYouMean(p)),a.innerHTML=this.renderPagination(),this.renderProductViewTypeUI(),this.renderBannerUI(),e(this,h)};var Wt=function(){var t,e,n;n=document,Math.max(n.body.scrollHeight,n.documentElement.scrollHeight,n.body.offsetHeight,n.documentElement.offsetHeight,n.body.clientHeight,n.documentElement.clientHeight)-20<=(t=0,e=0,"number"==typeof window.pageYOffset?(e=window.pageYOffset,t=window.pageXOffset):document.body&&(document.body.scrollLeft||document.body.scrollTop)?(e=document.body.scrollTop,t=document.body.scrollLeft):document.documentElement&&(document.documentElement.scrollLeft||document.documentElement.scrollTop)&&(e=document.documentElement.scrollTop,t=document.documentElement.scrollLeft),[t,e])[1]+window.innerHeight&&(this.viewState.isInfiniteStarted||(this.viewState.isInfiniteStarted=!0,this.renderNewResults("next")))},Ht=function(t,e){var n;return function(){var i=this,s=arguments;clearTimeout(n),n=setTimeout((function(){return t.apply(i,s)}),e)}};var Xt=function(){var t=this,e=this.options,n=e.paginationContainer,i=e.paginationEvt,s=e.sortContainer,a=e.searchButtonSelector,r=e.searchEvt,o=e.productItemClass,c=e.sortElem,l=e.sortEvt,h=e.facetEvt,u=e.facetElemWrapClass,d=e.selectedFacetBlock,p=e.selectedFacetClass,f=e.productViewTypeSelector;n.addEventListener(i,this.paginationAction.bind(this)),a.addEventListener(r,this.setInputValue.bind(this)),this.delegate(this.searchResultsWrapper,"click",o,this.onProductItemClick.bind(this)),this.delegate(s,l,c,this.sortAction.bind(this)),this.delegate(this.facetsWrapper,h,"."+u,this.findChangedFacet.bind(this)),d&&this.delegate(d,h,"."+p,this.findChangedFacet.bind(this)),this.rangeFacetsWrapper&&this.delegate(this.rangeFacetsWrapper,"click","button",this.onClickRangeFacet.bind(this)),this.bucketedSearchWrapper&&this.delegate(this.bucketedSearchWrapper,this.options.bucketFacetEvnt,"."+this.options.bucketedFacetElem,this.onBucketedFacet.bind(this)),this.breadcrumbWrapper&&this.delegate(this.breadcrumbWrapper,"click","."+this.options.breadcrumbSelectorClass,this.onBucketedFacet.bind(this)),f&&this.delegate(f,"click","button",this.onPageViewTypeClick.bind(this)),"INFINITE_SCROLL"===this.options.paginationType&&document.addEventListener("scroll",Ht((function(){Wt.bind(t)()}),1e3))},Yt=function(){this.renderDidYouMean.bind(this),this.options.callBackFn.bind(this),this.options.productClick.bind(this),this.callBack=this.callBack.bind(this),this.renderSwatchBtns=this.renderSwatchBtns.bind(this),this.reRender=Dt.bind(this),this.renderProductViewTypeUI(),this.getQueryParams()[this.options.searchQueryParam]&&this.renderFromUrl(),this.events=it,this.actions=st,this.bindEvents=Xt.bind(this)()};function qt(t){return(qt="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function Gt(t,e){for(var n=0;n<e.length;n++){var i=e[n];i.enumerable=i.enumerable||!1,i.configurable=!0,"value"in i&&(i.writable=!0),Object.defineProperty(t,i.key,i)}}function Qt(t,e){return(Qt=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function zt(t){var e=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(t){return!1}}();return function(){var n,i=$t(t);if(e){var s=$t(this).constructor;n=Reflect.construct(i,arguments,s)}else n=i.apply(this,arguments);return Kt(this,n)}}function Kt(t,e){return!e||"object"!==qt(e)&&"function"!=typeof e?Jt(t):e}function Jt(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function $t(t){return($t=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}var Zt=function(t){!function(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&Qt(t,e)}(a,t);var e,n,i,s=zt(a);function a(t){var e;return function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,a),(e=s.call(this,t)).options=Object.assign({},pt,t),e.viewState={productViewType:e.options.productViewTypes,isInfiniteStarted:!1},gt.bind(Jt(e))(),Yt.bind(Jt(e))(),e.widgets={RangeSlider:et.bind(Jt(e))},e}return e=a,(n=[{key:"callBack",value:function(t,e){var n=this.options,i=n.callBackFn,s=n.loaderElem,a=it.beforeApiCall,r=it.afterApiCall;e===a&&(i(this,a),this.loaderContainer.innerHTML=s(this)),e===r&&(i(this,r),this.reRender())}},{key:"delegate",value:function(t,e,n,i){return J(t,e,n,i)}}])&&Gt(e.prototype,n),i&&Gt(e,i),a}(K);_t(Zt);e.default=Zt}]);
