+++ src/core/setMethods.js
@@ -1,0 +1,26 @@
+import renderBannerUI from '../modules/banners/setbannerMethods';
+import setFacets from '../modules/facets/setFacetMethods';
+import setSearch from '../modules/searchResults/setSearch';
+import setProductViewType from '../modules/productViewType/setProductViewType';
+import setPagination from '../modules/pagination/setPagination';
+import setSort from '../modules/sort/setSort';
+import setInput from '../modules/input/setInput';
+import setSpellCheck from '../modules/didYouMean/setSpellCheck';
+import setBreadCrumbs from '../modules/breadcrumbs/setBreadcrumbs';
+import setSwatches from '../modules/swatches/setSwatches';
+const setMethods = (UnbxdSearch) => {
+    const {
+        prototype
+    } = UnbxdSearch;
+    prototype.renderBannerUI = renderBannerUI;
+    setInput(prototype);
+    setProductViewType(prototype);
+    setFacets(prototype);
+    setSearch(prototype);
+    setPagination(prototype);
+    setSort(prototype);
+    setSpellCheck(prototype);
+    setBreadCrumbs(prototype);
+    setSwatches(prototype);
+}
+export default setMethods;
