+++ src/modules/banners/setbannerMethods.js
@@ -1,0 +1,5 @@
+function renderBannerUI(){
+    const banners = this.getBanners();
+    this.bannerWrapper.innerHTML = this.options.bannerTemplate(banners);
+}
+export default renderBannerUI;
