+++ src/modules/breadcrumbs/setBreadcrumbs.js
@@ -1,0 +1,17 @@
+import {
+    onBreadCrumbClick,
+    renderBreadCrumbs
+} from './actions';
+
+const setBreadCrumbs = (prototype) => {
+    prototype = Object.assign(prototype,{
+        onBreadCrumbClick,
+        renderBreadCrumbs
+    })
+}
+
+export {
+    setBreadCrumbs as default,
+    onBreadCrumbClick,
+    renderBreadCrumbs
+};
