+++ src/modules/didYouMean/renderSpellCheck.js
@@ -1,0 +1,4 @@
+const renderDidYouMean = function (suggestion) {
+    return this.options.spellCheckTemplate.bind(this)(suggestion);
+}
+export default renderDidYouMean;
