+++ src/modules/didYouMean/setSpellCheck.js
@@ -1,0 +1,12 @@
+import renderDidYouMean from './renderSpellCheck';
+
+const setSpellCheck = (prototype) => {
+    prototype = Object.assign(prototype,{
+        renderDidYouMean
+    })
+}
+
+export {
+    setSpellCheck as default,
+    renderDidYouMean
+};
