+++ src/modules/didYouMean/spellCheckView.js
