+++ src/modules/facets/renderRangeFacets.js
