+++ src/modules/input/setInput.js
@@ -1,0 +1,12 @@
+import setInputValue from './actions';
+
+const setInput = (prototype) => {
+    prototype = Object.assign(prototype,{
+        setInputValue
+    })
+}
+
+export {
+    setInput as default,
+    setInputValue
+};
