+++ src/modules/searchResults/noResultsView.js
@@ -1,0 +1,4 @@
+function renderNoResults(query) {
+    return this.options.noResultsUi.bind(this)(query);
+}
+export default renderNoResults;
