+++ src/modules/sort/renderSort.js
@@ -1,0 +1,4 @@
+function renderSort(){
+    return this.options.sortUI.bind(this)(this.getSelectedSort());
+}
+export default renderSort;
