+++ src/modules/sort/setSort.js
@@ -1,0 +1,15 @@
+import sortAction from './actions';
+import renderSort from './renderSort';
+
+const setSort = (prototype) => {
+    prototype = Object.assign(prototype,{
+        sortAction,
+        renderSort
+    })
+}
+
+export {
+    setSort as default,
+    sortAction,
+    renderSort
+};
