+++ src/modules/swatches/actions.js
@@ -1,0 +1,7 @@
+function renderSwatchBtns(product) {
+    const swatchData = this.getSwatches(product, this.options.swatchMap);
+    return this.options.swatchTemplate(swatchData);
+};
+export {
+    renderSwatchBtns
+};
