+++ src/modules/swatches/setSwatches.js
@@ -1,0 +1,12 @@
+import {
+    renderSwatchBtns
+ } from './actions';
+const setSwatches = (prototype) => {
+    prototype = Object.assign(prototype,{
+        renderSwatchBtns
+    })
+}
+export {
+    setSwatches as default,
+    renderSwatchBtns
+};
